Refactor file architecture for better separation of concerns
Split engine.go (4,723 LOC -> ~570 LOC) into focused transaction handler files:
- account_root.go: AccountRoot type and serialization utilities
- apply_account.go: Account management transactions (SetRegularKey, SignerListSet, etc.)
- apply_escrow.go: Escrow transactions (Create, Finish, Cancel)
- apply_paychannel.go: Payment channel transactions
- apply_check.go: Check transactions
- apply_nftoken.go: NFToken transactions
- apply_amm.go: AMM transactions
- apply_misc.go: XChain, DID, Oracle, MPToken, Credential, Vault, Batch transactions

Split ledger service.go (1,958 LOC -> ~548 LOC) into focused files:
- account_query.go: Account-related queries (GetAccountInfo, GetAccountLines, etc.)
- ledger_query.go: Ledger queries (GetLedgerEntry, GetLedgerData, GetLedgerRange)
- tx_query.go: Transaction queries (GetTransaction, GetAccountTransactions)
- offer_query.go: Order book queries (GetBookOffers)
- persistence.go: Storage persistence logic
- helpers.go: Utility functions

Removed duplicate constant declarations across files.
