+++ internal/core/ledger/service/helpers.go
@@ -1,0 +1,178 @@
+package service
+
+import (
+	"errors"
+	"strconv"
+
+	addresscodec "github.com/LeJamon/goXRPLd/internal/codec/address-codec"
+	"github.com/LeJamon/goXRPLd/internal/core/tx"
+)
+
+// formatHashHex formats a hash as hex string
+func formatHashHex(hash [32]byte) string {
+	const hexChars = "0123456789ABCDEF"
+	result := make([]byte, 64)
+	for i, b := range hash {
+		result[i*2] = hexChars[b>>4]
+		result[i*2+1] = hexChars[b&0x0F]
+	}
+	return string(result)
+}
+
+// hexDecode decodes a hex string to bytes
+func hexDecode(s string) ([]byte, error) {
+	if len(s)%2 != 0 {
+		return nil, errors.New("odd length hex string")
+	}
+	result := make([]byte, len(s)/2)
+	for i := 0; i < len(s); i += 2 {
+		var b byte
+		for j := 0; j < 2; j++ {
+			c := s[i+j]
+			switch {
+			case c >= '0' && c <= '9':
+				b = b<<4 | (c - '0')
+			case c >= 'a' && c <= 'f':
+				b = b<<4 | (c - 'a' + 10)
+			case c >= 'A' && c <= 'F':
+				b = b<<4 | (c - 'A' + 10)
+			default:
+				return nil, errors.New("invalid hex character")
+			}
+		}
+		result[i/2] = b
+	}
+	return result, nil
+}
+
+// decodeAccountIDLocal decodes an account address to its 20-byte ID
+func decodeAccountIDLocal(address string) ([20]byte, error) {
+	var accountID [20]byte
+	if address == "" {
+		return accountID, errors.New("empty address")
+	}
+	_, accountIDBytes, err := addresscodec.DecodeClassicAddressToAccountID(address)
+	if err != nil {
+		return accountID, err
+	}
+	copy(accountID[:], accountIDBytes)
+	return accountID, nil
+}
+
+// amountsMatchCurrency checks if two amounts have the same currency (ignoring value)
+func amountsMatchCurrency(a, b tx.Amount) bool {
+	if a.IsNative() && b.IsNative() {
+		return true
+	}
+	if a.IsNative() != b.IsNative() {
+		return false
+	}
+	return a.Currency == b.Currency && a.Issuer == b.Issuer
+}
+
+// calculateOfferQuality calculates the quality (price) of an offer
+func calculateOfferQuality(pays, gets tx.Amount) string {
+	// Quality = TakerPays / TakerGets
+	paysVal := parseAmountValue(pays)
+	getsVal := parseAmountValue(gets)
+	if getsVal == 0 {
+		return "0"
+	}
+	quality := paysVal / getsVal
+	return strconv.FormatFloat(quality, 'g', -1, 64)
+}
+
+// parseAmountValue parses an amount value as float
+func parseAmountValue(amt tx.Amount) float64 {
+	if amt.IsNative() {
+		drops, _ := strconv.ParseUint(amt.Value, 10, 64)
+		return float64(drops)
+	}
+	val, _ := strconv.ParseFloat(amt.Value, 64)
+	return val
+}
+
+// formatHash formats a hash as a string
+func formatHash(hash [32]byte) string {
+	return string(hash[:])
+}
+
+// sortBookOffersByQuality sorts book offers by quality (best first)
+func sortBookOffersByQuality(offers []BookOffer) {
+	// Simple bubble sort - could use sort.Slice for better performance
+	for i := 0; i < len(offers)-1; i++ {
+		for j := i + 1; j < len(offers); j++ {
+			qi, _ := strconv.ParseFloat(offers[i].Quality, 64)
+			qj, _ := strconv.ParseFloat(offers[j].Quality, 64)
+			if qj < qi { // Lower quality is better (cheaper)
+				offers[i], offers[j] = offers[j], offers[i]
+			}
+		}
+	}
+}
+
+// helper function to format ledger range
+func formatRange(min, max uint32) string {
+	// Simple implementation - could be improved
+	return string(rune(min)) + "-" + string(rune(max))
+}
+
+// getLedgerEntryType extracts the entry type from serialized data
+func getLedgerEntryType(data []byte) string {
+	if len(data) < 3 {
+		return ""
+	}
+	if data[0] != 0x11 { // UInt16 type code
+		return ""
+	}
+	entryType := uint16(data[1])<<8 | uint16(data[2])
+	switch entryType {
+	case 0x0061: // 'a' = AccountRoot
+		return "AccountRoot"
+	case 0x0063: // 'c' = Check
+		return "Check"
+	case 0x0064: // 'd' = DirNode
+		return "DirectoryNode"
+	case 0x0066: // 'f' = FeeSettings
+		return "FeeSettings"
+	case 0x0068: // 'h' = Escrow
+		return "Escrow"
+	case 0x006E: // 'n' = NFTokenPage
+		return "NFTokenPage"
+	case 0x006F: // 'o' = Offer
+		return "Offer"
+	case 0x0070: // 'p' = PayChannel
+		return "PayChannel"
+	case 0x0072: // 'r' = RippleState
+		return "RippleState"
+	case 0x0073: // 's' = SignerList
+		return "SignerList"
+	case 0x0074: // 't' = Ticket
+		return "Ticket"
+	case 0x0075: // 'u' = NFTokenOffer
+		return "NFTokenOffer"
+	case 0x0078: // 'x' = AMM
+		return "AMM"
+	default:
+		return ""
+	}
+}
+
+// isObjectForAccount checks if a ledger object belongs to an account
+func isObjectForAccount(data []byte, accountID [20]byte, entryType string) bool {
+	// This is a simplified check - in production, properly parse the object
+	// For now, check if the account ID appears in the data
+	for i := 0; i <= len(data)-20; i++ {
+		match := true
+		for j := 0; j < 20; j++ {
+			if data[i+j] != accountID[j] {
+				match = false
+				break
+			}
+		}
+		if match {
+			return true
+		}
+	}
+	return false
+}
