+++ internal/core/tx/engine.go
@@ -1,14 +1,11 @@
 package tx
 
 import (
-	"encoding/binary"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"strconv"
 	"strings"
 
-	addresscodec "github.com/LeJamon/goXRPLd/internal/codec/address-codec"
 	binarycodec "github.com/LeJamon/goXRPLd/internal/codec/binary-codec"
 	"github.com/LeJamon/goXRPLd/internal/core/XRPAmount"
 	"github.com/LeJamon/goXRPLd/internal/core/ledger/keylet"
@@ -570,4154 +567,4 @@ func (e *Engine) calculateFee(tx Transaction) uint64 {
 		}
 	}
 	return e.config.BaseFee
-}
-
-// AccountRoot represents an account in the ledger
-type AccountRoot struct {
-	Account           string
-	Balance           uint64
-	Sequence          uint32
-	OwnerCount        uint32
-	Flags             uint32
-	RegularKey        string
-	Domain            string
-	EmailHash         string
-	MessageKey        string
-	TransferRate      uint32
-	TickSize          uint8
-	PreviousTxnID     [32]byte
-	PreviousTxnLgrSeq uint32
-}
-
-// AccountRoot binary format field codes (from XRPL spec)
-const (
-	// Field type codes
-	fieldTypeUInt16    = 1
-	fieldTypeUInt32    = 2
-	fieldTypeUInt64    = 3
-	fieldTypeHash128   = 4
-	fieldTypeHash256   = 5
-	fieldTypeAmount    = 6
-	fieldTypeBlob      = 7
-	fieldTypeAccount   = 8
-	fieldTypeAccountID = 8 // Same as Account, used in serialization
-
-	// Field codes for AccountRoot
-	fieldCodeLedgerEntryType = 1  // UInt16
-	fieldCodeFlags           = 2  // UInt32
-	fieldCodeSequence        = 4  // UInt32
-	fieldCodeOwnerCount      = 13 // UInt32 (per rippled sfields.macro)
-	fieldCodeTransferRate    = 11 // UInt32
-	fieldCodeBalance         = 1  // Amount
-	fieldCodeRegularKey      = 8  // Account
-	fieldCodeAccount         = 1  // Account (different context)
-	fieldCodeEmailHash       = 1  // Hash128
-	fieldCodeDomain          = 7  // Blob
-	fieldCodeTickSize        = 16 // UInt8 (stored as UInt16)
-
-	// Ledger entry type code for AccountRoot
-	ledgerEntryTypeAccountRoot = 0x0061
-)
-
-// AccountRoot ledger entry flags
-const (
-	// lsfDepositAuth indicates the account requires deposit authorization
-	// Payments to this account require preauthorization unless both the
-	// destination balance and payment amount are at or below the base reserve
-	lsfDepositAuth uint32 = 0x01000000
-
-	// lsfAMM indicates the account is an AMM (pseudo-account)
-	// AMM accounts cannot receive direct payments
-	lsfAMM uint32 = 0x02000000
-)
-
-// Helper functions
-
-func decodeAccountID(address string) ([20]byte, error) {
-	var accountID [20]byte
-	if address == "" {
-		return accountID, errors.New("empty address")
-	}
-
-	// Use the address codec to decode
-	_, accountIDBytes, err := addresscodec.DecodeClassicAddressToAccountID(address)
-	if err != nil {
-		return accountID, errors.New("invalid address: " + err.Error())
-	}
-
-	copy(accountID[:], accountIDBytes)
-	return accountID, nil
-}
-
-func encodeAccountID(accountID [20]byte) (string, error) {
-	return addresscodec.EncodeAccountIDToClassicAddress(accountID[:])
-}
-
-// ParseAccountRootFromBytes parses account data from binary format
-func ParseAccountRootFromBytes(data []byte) (*AccountRoot, error) {
-	return parseAccountRoot(data)
-}
-
-func parseAccountRoot(data []byte) (*AccountRoot, error) {
-	if len(data) < 20 {
-		return nil, errors.New("account data too short")
-	}
-
-	account := &AccountRoot{}
-	ledgerEntryTypeVerified := false // Track if we've verified the LedgerEntryType
-
-	// Parse the binary format
-	// XRPL uses a TLV-like format with field headers
-	offset := 0
-
-	for offset < len(data) {
-		if offset+1 > len(data) {
-			break
-		}
-
-		// Read field header
-		header := data[offset]
-		offset++
-
-		// Decode type and field from header
-		typeCode := (header >> 4) & 0x0F
-		fieldCode := header & 0x0F
-
-		// Handle extended type codes
-		if typeCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			typeCode = data[offset]
-			offset++
-		}
-
-		// Handle extended field codes
-		if fieldCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			fieldCode = data[offset]
-			offset++
-		}
-
-		// Parse field based on type
-		switch typeCode {
-		case fieldTypeUInt16:
-			if offset+2 > len(data) {
-				return account, nil
-			}
-			value := binary.BigEndian.Uint16(data[offset : offset+2])
-			offset += 2
-			// Only check LedgerEntryType once (first UInt16 with fieldCode 1)
-			if fieldCode == fieldCodeLedgerEntryType && !ledgerEntryTypeVerified {
-				// LedgerEntryType - verify it's AccountRoot
-				if value != ledgerEntryTypeAccountRoot {
-					return nil, errors.New("not an AccountRoot entry")
-				}
-				ledgerEntryTypeVerified = true
-			}
-
-		case fieldTypeUInt32:
-			if offset+4 > len(data) {
-				return account, nil
-			}
-			value := binary.BigEndian.Uint32(data[offset : offset+4])
-			offset += 4
-			switch fieldCode {
-			case fieldCodeFlags:
-				account.Flags = value
-			case fieldCodeSequence:
-				account.Sequence = value
-			case 5: // PreviousTxnLgrSeq
-				account.PreviousTxnLgrSeq = value
-			case fieldCodeOwnerCount:
-				account.OwnerCount = value
-			case fieldCodeTransferRate:
-				account.TransferRate = value
-			}
-
-		case fieldTypeAmount:
-			// XRP amounts are 8 bytes, IOU amounts are 48 bytes
-			if offset+8 > len(data) {
-				return account, nil
-			}
-			// Check if it's XRP (first bit is 0) or IOU (first bit is 1)
-			if data[offset]&0x80 == 0 {
-				// XRP amount - 8 bytes
-				// The format is: top bit = 0 for XRP, next bit = positive, remaining 62 bits = drops
-				rawAmount := binary.BigEndian.Uint64(data[offset : offset+8])
-				// Clear the top bit and extract drops
-				account.Balance = rawAmount & 0x3FFFFFFFFFFFFFFF
-				offset += 8
-			} else {
-				// IOU amount - skip 48 bytes (we don't handle this in AccountRoot)
-				offset += 48
-			}
-
-		case fieldTypeAccount:
-			// Account IDs are variable length encoded
-			if offset >= len(data) {
-				return account, nil
-			}
-			length := int(data[offset])
-			offset++
-			if offset+length > len(data) {
-				return account, nil
-			}
-			if length == 20 {
-				var accID [20]byte
-				copy(accID[:], data[offset:offset+length])
-				addr, err := encodeAccountID(accID)
-				if err == nil {
-					if fieldCode == fieldCodeAccount || fieldCode == 1 {
-						account.Account = addr
-					} else if fieldCode == fieldCodeRegularKey {
-						account.RegularKey = addr
-					}
-				}
-			}
-			offset += length
-
-		case fieldTypeBlob:
-			// Variable length blob
-			if offset >= len(data) {
-				return account, nil
-			}
-			length := int(data[offset])
-			offset++
-			if length > 192 {
-				// Extended length encoding
-				if length < 241 {
-					length = 193 + int(data[offset-1]) - 193
-				} else {
-					// Even more extended - skip for now
-					offset += 2
-					continue
-				}
-			}
-			if offset+length > len(data) {
-				return account, nil
-			}
-			if fieldCode == 7 { // Domain field
-				account.Domain = string(data[offset : offset+length])
-			}
-			offset += length
-
-		case fieldTypeHash128:
-			if offset+16 > len(data) {
-				return account, nil
-			}
-			if fieldCode == fieldCodeEmailHash {
-				account.EmailHash = hex.EncodeToString(data[offset : offset+16])
-			}
-			offset += 16
-
-		case fieldTypeHash256:
-			// Hash256 fields (e.g., PreviousTxnID) are 32 bytes
-			if offset+32 > len(data) {
-				return account, nil
-			}
-			if fieldCode == 5 { // PreviousTxnID
-				copy(account.PreviousTxnID[:], data[offset:offset+32])
-			}
-			offset += 32
-
-		default:
-			// Unknown type - can't determine size, must stop parsing
-			// This shouldn't happen for valid AccountRoot entries
-			return account, nil
-		}
-	}
-
-	return account, nil
-}
-
-func serializeAccountRoot(account *AccountRoot) ([]byte, error) {
-	// Build the JSON representation for the binary codec
-	jsonObj := map[string]any{
-		"LedgerEntryType": "AccountRoot",
-		"Balance":         fmt.Sprintf("%d", account.Balance), // XRP balance as drops string
-		"Sequence":        account.Sequence,
-		"OwnerCount":      account.OwnerCount,
-		"Flags":           account.Flags,
-	}
-
-	// Add Account if set
-	if account.Account != "" {
-		jsonObj["Account"] = account.Account
-	}
-
-	// Add TransferRate if set
-	if account.TransferRate > 0 {
-		jsonObj["TransferRate"] = account.TransferRate
-	}
-
-	// Add RegularKey if set
-	if account.RegularKey != "" {
-		jsonObj["RegularKey"] = account.RegularKey
-	}
-
-	// Add Domain if set (as hex string)
-	if account.Domain != "" {
-		jsonObj["Domain"] = strings.ToUpper(hex.EncodeToString([]byte(account.Domain)))
-	}
-
-	// Add EmailHash if set
-	if account.EmailHash != "" {
-		jsonObj["EmailHash"] = strings.ToUpper(account.EmailHash)
-	}
-
-	// Add PreviousTxnID if set (non-zero)
-	var zeroHash [32]byte
-	if account.PreviousTxnID != zeroHash {
-		jsonObj["PreviousTxnID"] = strings.ToUpper(hex.EncodeToString(account.PreviousTxnID[:]))
-	}
-
-	// Add PreviousTxnLgrSeq if set
-	if account.PreviousTxnLgrSeq > 0 {
-		jsonObj["PreviousTxnLgrSeq"] = account.PreviousTxnLgrSeq
-	}
-
-	// Encode using the binary codec
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode AccountRoot: %w", err)
-	}
-
-	// Convert hex string to bytes
-	return hex.DecodeString(hexStr)
-}
-
-// applySetRegularKey applies a SetRegularKey transaction
-func (e *Engine) applySetRegularKey(tx *SetRegularKey, account *AccountRoot, metadata *Metadata) Result {
-	// Update the account's regular key
-	previousRegularKey := account.RegularKey
-	account.RegularKey = tx.RegularKey
-
-	// If setting a new key, validate it exists (or just validate format)
-	if tx.RegularKey != "" {
-		if _, err := decodeAccountID(tx.RegularKey); err != nil {
-			return TemINVALID
-		}
-	}
-
-	// Record the change in metadata
-	if previousRegularKey != account.RegularKey {
-		// This will be recorded in the main account update
-	}
-
-	return TesSUCCESS
-}
-
-// applySignerListSet applies a SignerListSet transaction
-func (e *Engine) applySignerListSet(tx *SignerListSet, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Create the SignerList keylet
-	signerListKey := keylet.SignerList(accountID)
-
-	if tx.SignerQuorum == 0 {
-		// Delete existing signer list
-		exists, _ := e.view.Exists(signerListKey)
-		if exists {
-			if err := e.view.Erase(signerListKey); err != nil {
-				return TefINTERNAL
-			}
-
-			// Decrease owner count
-			if account.OwnerCount > 0 {
-				account.OwnerCount--
-			}
-
-			// Record deletion in metadata
-			metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-				NodeType:        "DeletedNode",
-				LedgerEntryType: "SignerList",
-				LedgerIndex:     hex.EncodeToString(signerListKey.Key[:]),
-			})
-		}
-	} else {
-		// Create or update signer list
-		signerListData, err := serializeSignerList(tx, accountID)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		exists, _ := e.view.Exists(signerListKey)
-		if exists {
-			// Update existing
-			if err := e.view.Update(signerListKey, signerListData); err != nil {
-				return TefINTERNAL
-			}
-			metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-				NodeType:        "ModifiedNode",
-				LedgerEntryType: "SignerList",
-				LedgerIndex:     hex.EncodeToString(signerListKey.Key[:]),
-			})
-		} else {
-			// Create new
-			if err := e.view.Insert(signerListKey, signerListData); err != nil {
-				return TefINTERNAL
-			}
-
-			// Increase owner count
-			account.OwnerCount++
-
-			metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-				NodeType:        "CreatedNode",
-				LedgerEntryType: "SignerList",
-				LedgerIndex:     hex.EncodeToString(signerListKey.Key[:]),
-			})
-		}
-	}
-
-	return TesSUCCESS
-}
-
-// applyTicketCreate applies a TicketCreate transaction
-func (e *Engine) applyTicketCreate(tx *TicketCreate, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Create tickets
-	for i := uint32(0); i < tx.TicketCount; i++ {
-		ticketSeq := account.Sequence + i
-
-		// Create ticket keylet
-		ticketKey := keylet.Ticket(accountID, ticketSeq)
-
-		// Serialize ticket
-		ticketData, err := serializeTicket(accountID, ticketSeq)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		// Insert ticket
-		if err := e.view.Insert(ticketKey, ticketData); err != nil {
-			return TefINTERNAL
-		}
-
-		// Record creation in metadata
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "Ticket",
-			LedgerIndex:     hex.EncodeToString(ticketKey.Key[:]),
-			NewFields: map[string]any{
-				"Account":        tx.Account,
-				"TicketSequence": ticketSeq,
-			},
-		})
-	}
-
-	// Increase owner count for each ticket
-	account.OwnerCount += tx.TicketCount
-
-	// Increase sequence by ticket count (tickets consume sequence numbers)
-	account.Sequence += tx.TicketCount
-
-	return TesSUCCESS
-}
-
-// applyDepositPreauth applies a DepositPreauth transaction
-func (e *Engine) applyDepositPreauth(tx *DepositPreauth, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	if tx.Authorize != "" {
-		// Create preauthorization
-		authorizedID, err := decodeAccountID(tx.Authorize)
-		if err != nil {
-			return TemINVALID
-		}
-
-		// Check that authorized account exists
-		authorizedKey := keylet.Account(authorizedID)
-		exists, _ := e.view.Exists(authorizedKey)
-		if !exists {
-			return TecNO_TARGET
-		}
-
-		// Create deposit preauth keylet
-		preauthKey := keylet.DepositPreauth(accountID, authorizedID)
-
-		// Check if already exists
-		exists, _ = e.view.Exists(preauthKey)
-		if exists {
-			return TecDUPLICATE
-		}
-
-		// Serialize and insert
-		preauthData, err := serializeDepositPreauth(accountID, authorizedID)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		if err := e.view.Insert(preauthKey, preauthData); err != nil {
-			return TefINTERNAL
-		}
-
-		// Increase owner count
-		account.OwnerCount++
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "DepositPreauth",
-			LedgerIndex:     hex.EncodeToString(preauthKey.Key[:]),
-			NewFields: map[string]any{
-				"Account":   tx.Account,
-				"Authorize": tx.Authorize,
-			},
-		})
-	} else if tx.Unauthorize != "" {
-		// Remove preauthorization
-		unauthorizedID, err := decodeAccountID(tx.Unauthorize)
-		if err != nil {
-			return TemINVALID
-		}
-
-		preauthKey := keylet.DepositPreauth(accountID, unauthorizedID)
-
-		// Check if exists
-		exists, _ := e.view.Exists(preauthKey)
-		if !exists {
-			return TecNO_ENTRY
-		}
-
-		// Delete
-		if err := e.view.Erase(preauthKey); err != nil {
-			return TefINTERNAL
-		}
-
-		// Decrease owner count
-		if account.OwnerCount > 0 {
-			account.OwnerCount--
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "DepositPreauth",
-			LedgerIndex:     hex.EncodeToString(preauthKey.Key[:]),
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// applyAccountDelete applies an AccountDelete transaction
-func (e *Engine) applyAccountDelete(tx *AccountDelete, account *AccountRoot, metadata *Metadata) Result {
-	// Check that owner count is 0 (no objects owned)
-	if account.OwnerCount > 0 {
-		return TecHAS_OBLIGATIONS
-	}
-
-	// Check minimum sequence requirement (account must have been around for a while)
-	// In standalone mode, we relax this requirement
-	if !e.config.Standalone && account.Sequence < 256 {
-		return TefTOO_BIG // Account too young
-	}
-
-	// Get destination account
-	destID, err := decodeAccountID(tx.Destination)
-	if err != nil {
-		return TemINVALID
-	}
-
-	destKey := keylet.Account(destID)
-	destData, err := e.view.Read(destKey)
-	if err != nil {
-		return TecNO_DST
-	}
-
-	destAccount, err := parseAccountRoot(destData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Calculate remaining balance (after fee was deducted)
-	remainingBalance := account.Balance
-
-	// Transfer remaining balance to destination
-	destAccount.Balance += remainingBalance
-
-	// Update destination account
-	destUpdatedData, err := serializeAccountRoot(destAccount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Update(destKey, destUpdatedData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Delete source account
-	srcID, _ := decodeAccountID(tx.Account)
-	srcKey := keylet.Account(srcID)
-
-	if err := e.view.Erase(srcKey); err != nil {
-		return TefINTERNAL
-	}
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "AccountRoot",
-		LedgerIndex:     hex.EncodeToString(srcKey.Key[:]),
-		FinalFields: map[string]any{
-			"Account": tx.Account,
-			"Balance": "0",
-		},
-	})
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AccountRoot",
-		LedgerIndex:     hex.EncodeToString(destKey.Key[:]),
-		FinalFields: map[string]any{
-			"Account": tx.Destination,
-			"Balance": strconv.FormatUint(destAccount.Balance, 10),
-		},
-	})
-
-	// Set account balance to 0 so the main update doesn't try to write it
-	account.Balance = 0
-
-	return TesSUCCESS
-}
-
-// Helper function to serialize a SignerList
-func serializeSignerList(tx *SignerListSet, ownerID [20]byte) ([]byte, error) {
-	// Convert owner ID to classic address
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	// Build the JSON representation for the binary codec
-	jsonObj := map[string]any{
-		"LedgerEntryType": "SignerList",
-		"Account":         ownerAddress,
-		"SignerQuorum":    tx.SignerQuorum,
-		"OwnerNode":       "0", // UInt64 as string
-	}
-
-	// Add SignerEntries if present
-	if len(tx.SignerEntries) > 0 {
-		signerEntries := make([]map[string]any, len(tx.SignerEntries))
-		for i, entry := range tx.SignerEntries {
-			signerEntries[i] = map[string]any{
-				"SignerEntry": map[string]any{
-					"Account":      entry.SignerEntry.Account,
-					"SignerWeight": entry.SignerEntry.SignerWeight,
-				},
-			}
-		}
-		jsonObj["SignerEntries"] = signerEntries
-	}
-
-	// Encode using the binary codec
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode SignerList: %w", err)
-	}
-
-	// Convert hex string to bytes
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to serialize a Ticket
-func serializeTicket(ownerID [20]byte, ticketSeq uint32) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "Ticket",
-		"Account":         ownerAddress,
-		"TicketSequence":  ticketSeq,
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode Ticket: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to serialize a DepositPreauth
-func serializeDepositPreauth(ownerID, authorizedID [20]byte) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	authorizedAddress, err := addresscodec.EncodeAccountIDToClassicAddress(authorizedID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode authorized address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "DepositPreauth",
-		"Account":         ownerAddress,
-		"Authorize":       authorizedAddress,
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode DepositPreauth: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// applyEscrowCreate applies an EscrowCreate transaction
-func (e *Engine) applyEscrowCreate(tx *EscrowCreate, account *AccountRoot, metadata *Metadata) Result {
-	// Parse the amount to escrow
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err != nil {
-		return TemINVALID
-	}
-
-	// Check that account has sufficient balance (after fee)
-	if account.Balance < amount {
-		return TecUNFUNDED
-	}
-
-	// Verify destination exists
-	destID, err := decodeAccountID(tx.Destination)
-	if err != nil {
-		return TemINVALID
-	}
-
-	destKey := keylet.Account(destID)
-	exists, _ := e.view.Exists(destKey)
-	if !exists {
-		return TecNO_DST
-	}
-
-	// Deduct the escrow amount from the account
-	account.Balance -= amount
-
-	// Create the escrow entry
-	accountID, _ := decodeAccountID(tx.Account)
-	sequence := *tx.GetCommon().Sequence // Use the transaction sequence
-
-	escrowKey := keylet.Escrow(accountID, sequence)
-
-	// Serialize escrow
-	escrowData, err := serializeEscrow(tx, accountID, destID, sequence, amount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Insert escrow
-	if err := e.view.Insert(escrowKey, escrowData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Increase owner count
-	account.OwnerCount++
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "Escrow",
-		LedgerIndex:     hex.EncodeToString(escrowKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":     tx.Account,
-			"Destination": tx.Destination,
-			"Amount":      tx.Amount.Value,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyEscrowFinish applies an EscrowFinish transaction
-func (e *Engine) applyEscrowFinish(tx *EscrowFinish, account *AccountRoot, metadata *Metadata) Result {
-	// Get the escrow owner's account ID
-	ownerID, err := decodeAccountID(tx.Owner)
-	if err != nil {
-		return TemINVALID
-	}
-
-	// Find the escrow
-	escrowKey := keylet.Escrow(ownerID, tx.OfferSequence)
-	escrowData, err := e.view.Read(escrowKey)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	// Parse escrow
-	escrow, err := parseEscrow(escrowData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Check FinishAfter time (if set)
-	if escrow.FinishAfter > 0 {
-		// In a full implementation, we'd check against the close time
-		// For now, we'll allow it in standalone mode
-		if !e.config.Standalone {
-			// Would check: if currentTime < escrow.FinishAfter return TecNO_PERMISSION
-		}
-	}
-
-	// Check condition/fulfillment (simplified - in reality, would verify crypto-condition)
-	if escrow.Condition != "" {
-		if tx.Fulfillment == "" {
-			return TecCRYPTOCONDITION_ERROR
-		}
-		// Would verify: fulfillment matches condition
-	}
-
-	// Get destination account
-	destKey := keylet.Account(escrow.DestinationID)
-	destData, err := e.view.Read(destKey)
-	if err != nil {
-		return TecNO_DST
-	}
-
-	destAccount, err := parseAccountRoot(destData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Transfer the escrowed amount to destination
-	destAccount.Balance += escrow.Amount
-
-	// Update destination
-	destUpdatedData, err := serializeAccountRoot(destAccount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Update(destKey, destUpdatedData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Delete the escrow
-	if err := e.view.Erase(escrowKey); err != nil {
-		return TefINTERNAL
-	}
-
-	// Decrease owner count for escrow owner
-	if tx.Owner != tx.Account {
-		// Need to update owner's account too
-		ownerKey := keylet.Account(ownerID)
-		ownerData, err := e.view.Read(ownerKey)
-		if err == nil {
-			ownerAccount, err := parseAccountRoot(ownerData)
-			if err == nil && ownerAccount.OwnerCount > 0 {
-				ownerAccount.OwnerCount--
-				ownerUpdatedData, err := serializeAccountRoot(ownerAccount)
-				if err == nil {
-					e.view.Update(ownerKey, ownerUpdatedData)
-				}
-			}
-		}
-	} else {
-		if account.OwnerCount > 0 {
-			account.OwnerCount--
-		}
-	}
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Escrow",
-		LedgerIndex:     hex.EncodeToString(escrowKey.Key[:]),
-	})
-
-	destAddr, _ := encodeAccountID(escrow.DestinationID)
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AccountRoot",
-		LedgerIndex:     hex.EncodeToString(destKey.Key[:]),
-		FinalFields: map[string]any{
-			"Account": destAddr,
-			"Balance": strconv.FormatUint(destAccount.Balance, 10),
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyEscrowCancel applies an EscrowCancel transaction
-func (e *Engine) applyEscrowCancel(tx *EscrowCancel, account *AccountRoot, metadata *Metadata) Result {
-	// Get the escrow owner's account ID
-	ownerID, err := decodeAccountID(tx.Owner)
-	if err != nil {
-		return TemINVALID
-	}
-
-	// Find the escrow
-	escrowKey := keylet.Escrow(ownerID, tx.OfferSequence)
-	escrowData, err := e.view.Read(escrowKey)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	// Parse escrow
-	escrow, err := parseEscrow(escrowData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Check CancelAfter time (if set)
-	if escrow.CancelAfter > 0 {
-		// In a full implementation, we'd check against the close time
-		// For now, we'll allow it in standalone mode
-		if !e.config.Standalone {
-			// Would check: if currentTime < escrow.CancelAfter return TecNO_PERMISSION
-		}
-	} else {
-		// If no CancelAfter, only the creator can cancel (implied by having condition)
-		if tx.Account != tx.Owner {
-			return TecNO_PERMISSION
-		}
-	}
-
-	// Return the escrowed amount to the owner
-	ownerKey := keylet.Account(ownerID)
-	ownerData, err := e.view.Read(ownerKey)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	ownerAccount, err := parseAccountRoot(ownerData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	ownerAccount.Balance += escrow.Amount
-	if ownerAccount.OwnerCount > 0 {
-		ownerAccount.OwnerCount--
-	}
-
-	ownerUpdatedData, err := serializeAccountRoot(ownerAccount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Update(ownerKey, ownerUpdatedData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Delete the escrow
-	if err := e.view.Erase(escrowKey); err != nil {
-		return TefINTERNAL
-	}
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Escrow",
-		LedgerIndex:     hex.EncodeToString(escrowKey.Key[:]),
-	})
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AccountRoot",
-		LedgerIndex:     hex.EncodeToString(ownerKey.Key[:]),
-		FinalFields: map[string]any{
-			"Account": tx.Owner,
-			"Balance": strconv.FormatUint(ownerAccount.Balance, 10),
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// Escrow ledger entry data
-type EscrowData struct {
-	Account       [20]byte
-	DestinationID [20]byte
-	Amount        uint64
-	Condition     string
-	CancelAfter   uint32
-	FinishAfter   uint32
-}
-
-// Helper function to serialize an Escrow
-func serializeEscrow(tx *EscrowCreate, ownerID, destID [20]byte, sequence uint32, amount uint64) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	destAddress, err := addresscodec.EncodeAccountIDToClassicAddress(destID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode destination address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "Escrow",
-		"Account":         ownerAddress,
-		"Destination":     destAddress,
-		"Amount":          fmt.Sprintf("%d", amount),
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	if tx.FinishAfter != nil {
-		jsonObj["FinishAfter"] = *tx.FinishAfter
-	}
-
-	if tx.CancelAfter != nil {
-		jsonObj["CancelAfter"] = *tx.CancelAfter
-	}
-
-	if tx.Condition != "" {
-		jsonObj["Condition"] = tx.Condition
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode Escrow: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to parse an Escrow ledger entry
-func parseEscrow(data []byte) (*EscrowData, error) {
-	escrow := &EscrowData{}
-	offset := 0
-
-	for offset < len(data) {
-		if offset+1 > len(data) {
-			break
-		}
-
-		header := data[offset]
-		offset++
-
-		typeCode := (header >> 4) & 0x0F
-		fieldCode := header & 0x0F
-
-		if typeCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			typeCode = data[offset]
-			offset++
-		}
-
-		if fieldCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			fieldCode = data[offset]
-			offset++
-		}
-
-		switch typeCode {
-		case fieldTypeUInt16:
-			if offset+2 > len(data) {
-				return escrow, nil
-			}
-			offset += 2
-
-		case fieldTypeUInt32:
-			if offset+4 > len(data) {
-				return escrow, nil
-			}
-			value := binary.BigEndian.Uint32(data[offset : offset+4])
-			offset += 4
-			switch fieldCode {
-			case 36: // FinishAfter
-				escrow.FinishAfter = value
-			case 37: // CancelAfter
-				escrow.CancelAfter = value
-			}
-
-		case fieldTypeUInt64:
-			if offset+8 > len(data) {
-				return escrow, nil
-			}
-			offset += 8
-
-		case fieldTypeAmount:
-			if offset+8 > len(data) {
-				return escrow, nil
-			}
-			rawAmount := binary.BigEndian.Uint64(data[offset : offset+8])
-			escrow.Amount = rawAmount & 0x3FFFFFFFFFFFFFFF
-			offset += 8
-
-		case fieldTypeAccountID:
-			if offset+21 > len(data) {
-				return escrow, nil
-			}
-			length := data[offset]
-			offset++
-			if length == 20 {
-				switch fieldCode {
-				case 1: // Account
-					copy(escrow.Account[:], data[offset:offset+20])
-				case 3: // Destination
-					copy(escrow.DestinationID[:], data[offset:offset+20])
-				}
-				offset += 20
-			}
-
-		case fieldTypeBlob:
-			if offset >= len(data) {
-				return escrow, nil
-			}
-			length := int(data[offset])
-			offset++
-			if offset+length > len(data) {
-				return escrow, nil
-			}
-			if fieldCode == 25 { // Condition
-				escrow.Condition = hex.EncodeToString(data[offset : offset+length])
-			}
-			offset += length
-
-		default:
-			// Unknown type - try to skip safely
-			return escrow, nil
-		}
-	}
-
-	return escrow, nil
-}
-
-// PayChannel ledger entry data
-type PayChannelData struct {
-	Account       [20]byte
-	DestinationID [20]byte
-	Amount        uint64
-	Balance       uint64
-	SettleDelay   uint32
-	PublicKey     string
-	Expiration    uint32
-	CancelAfter   uint32
-}
-
-// applyPaymentChannelCreate applies a PaymentChannelCreate transaction
-func (e *Engine) applyPaymentChannelCreate(tx *PaymentChannelCreate, account *AccountRoot, metadata *Metadata) Result {
-	// Parse the amount
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err != nil {
-		return TemINVALID
-	}
-
-	// Check balance
-	if account.Balance < amount {
-		return TecUNFUNDED
-	}
-
-	// Verify destination exists
-	destID, err := decodeAccountID(tx.Destination)
-	if err != nil {
-		return TemINVALID
-	}
-
-	destKey := keylet.Account(destID)
-	exists, _ := e.view.Exists(destKey)
-	if !exists {
-		return TecNO_DST
-	}
-
-	// Deduct amount from account
-	account.Balance -= amount
-
-	// Create pay channel
-	accountID, _ := decodeAccountID(tx.Account)
-	sequence := *tx.GetCommon().Sequence
-
-	channelKey := keylet.PayChannel(accountID, destID, sequence)
-
-	// Serialize pay channel
-	channelData, err := serializePayChannel(tx, accountID, destID, amount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Insert channel
-	if err := e.view.Insert(channelKey, channelData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Increase owner count
-	account.OwnerCount++
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "PayChannel",
-		LedgerIndex:     hex.EncodeToString(channelKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":     tx.Account,
-			"Destination": tx.Destination,
-			"Amount":      tx.Amount.Value,
-			"Balance":     "0",
-			"SettleDelay": tx.SettleDelay,
-			"PublicKey":   tx.PublicKey,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyPaymentChannelFund applies a PaymentChannelFund transaction
-func (e *Engine) applyPaymentChannelFund(tx *PaymentChannelFund, account *AccountRoot, metadata *Metadata) Result {
-	// Parse channel ID
-	channelID, err := hex.DecodeString(tx.Channel)
-	if err != nil || len(channelID) != 32 {
-		return TemINVALID
-	}
-
-	var channelKeyBytes [32]byte
-	copy(channelKeyBytes[:], channelID)
-	channelKey := keylet.Keylet{Key: channelKeyBytes}
-
-	// Read channel
-	channelData, err := e.view.Read(channelKey)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	// Parse channel
-	channel, err := parsePayChannel(channelData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Verify sender is the channel owner
-	accountID, _ := decodeAccountID(tx.Account)
-	if channel.Account != accountID {
-		return TecNO_PERMISSION
-	}
-
-	// Parse amount to add
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err != nil {
-		return TemINVALID
-	}
-
-	// Check balance
-	if account.Balance < amount {
-		return TecUNFUNDED
-	}
-
-	// Deduct from account
-	account.Balance -= amount
-
-	// Add to channel
-	channel.Amount += amount
-
-	// Update expiration if specified
-	if tx.Expiration != nil {
-		channel.Expiration = *tx.Expiration
-	}
-
-	// Serialize updated channel
-	updatedChannelData, err := serializePayChannelFromData(channel)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Update(channelKey, updatedChannelData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "PayChannel",
-		LedgerIndex:     hex.EncodeToString(channelKey.Key[:]),
-		FinalFields: map[string]any{
-			"Amount": strconv.FormatUint(channel.Amount, 10),
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyPaymentChannelClaim applies a PaymentChannelClaim transaction
-func (e *Engine) applyPaymentChannelClaim(tx *PaymentChannelClaim, account *AccountRoot, metadata *Metadata) Result {
-	// Parse channel ID
-	channelID, err := hex.DecodeString(tx.Channel)
-	if err != nil || len(channelID) != 32 {
-		return TemINVALID
-	}
-
-	var channelKeyBytes [32]byte
-	copy(channelKeyBytes[:], channelID)
-	channelKey := keylet.Keylet{Key: channelKeyBytes}
-
-	// Read channel
-	channelData, err := e.view.Read(channelKey)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	// Parse channel
-	channel, err := parsePayChannel(channelData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	accountID, _ := decodeAccountID(tx.Account)
-	isOwner := channel.Account == accountID
-	isDest := channel.DestinationID == accountID
-
-	if !isOwner && !isDest {
-		return TecNO_PERMISSION
-	}
-
-	// Handle claim with signature
-	if tx.Balance != nil && tx.Amount != nil && tx.Signature != "" {
-		// Parse claimed balance
-		claimBalance, err := strconv.ParseUint(tx.Balance.Value, 10, 64)
-		if err != nil {
-			return TemINVALID
-		}
-
-		// Verify claim is valid (would verify signature in full implementation)
-		if claimBalance > channel.Amount {
-			return TecUNFUNDED_PAYMENT
-		}
-
-		if claimBalance < channel.Balance {
-			return TemINVALID // Can't decrease balance
-		}
-
-		// Calculate amount to transfer
-		transferAmount := claimBalance - channel.Balance
-
-		// Transfer to destination
-		destKey := keylet.Account(channel.DestinationID)
-		destData, err := e.view.Read(destKey)
-		if err != nil {
-			return TecNO_DST
-		}
-
-		destAccount, err := parseAccountRoot(destData)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		destAccount.Balance += transferAmount
-		channel.Balance = claimBalance
-
-		destUpdatedData, err := serializeAccountRoot(destAccount)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		if err := e.view.Update(destKey, destUpdatedData); err != nil {
-			return TefINTERNAL
-		}
-
-		destAddr, _ := encodeAccountID(channel.DestinationID)
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "AccountRoot",
-			LedgerIndex:     hex.EncodeToString(destKey.Key[:]),
-			FinalFields: map[string]any{
-				"Account": destAddr,
-				"Balance": strconv.FormatUint(destAccount.Balance, 10),
-			},
-		})
-	}
-
-	// Handle close flag
-	flags := tx.GetFlags()
-	if flags&PaymentChannelClaimFlagClose != 0 {
-		// Close the channel
-
-		// Return remaining funds to owner
-		remaining := channel.Amount - channel.Balance
-		if remaining > 0 {
-			ownerKey := keylet.Account(channel.Account)
-			ownerData, err := e.view.Read(ownerKey)
-			if err == nil {
-				ownerAccount, err := parseAccountRoot(ownerData)
-				if err == nil {
-					ownerAccount.Balance += remaining
-					if ownerAccount.OwnerCount > 0 {
-						ownerAccount.OwnerCount--
-					}
-					ownerUpdatedData, _ := serializeAccountRoot(ownerAccount)
-					e.view.Update(ownerKey, ownerUpdatedData)
-				}
-			}
-		}
-
-		// Delete channel
-		if err := e.view.Erase(channelKey); err != nil {
-			return TefINTERNAL
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "PayChannel",
-			LedgerIndex:     hex.EncodeToString(channelKey.Key[:]),
-		})
-	} else {
-		// Update channel
-		updatedChannelData, err := serializePayChannelFromData(channel)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		if err := e.view.Update(channelKey, updatedChannelData); err != nil {
-			return TefINTERNAL
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "PayChannel",
-			LedgerIndex:     hex.EncodeToString(channelKey.Key[:]),
-			FinalFields: map[string]any{
-				"Balance": strconv.FormatUint(channel.Balance, 10),
-			},
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// Helper function to serialize a PayChannel
-func serializePayChannel(tx *PaymentChannelCreate, ownerID, destID [20]byte, amount uint64) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	destAddress, err := addresscodec.EncodeAccountIDToClassicAddress(destID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode destination address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "PayChannel",
-		"Account":         ownerAddress,
-		"Destination":     destAddress,
-		"Amount":          fmt.Sprintf("%d", amount),
-		"Balance":         "0",
-		"SettleDelay":     tx.SettleDelay,
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	if tx.CancelAfter != nil {
-		jsonObj["CancelAfter"] = *tx.CancelAfter
-	}
-
-	if tx.PublicKey != "" {
-		jsonObj["PublicKey"] = tx.PublicKey
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode PayChannel: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to serialize a PayChannel from data
-func serializePayChannelFromData(channel *PayChannelData) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(channel.Account[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	destAddress, err := addresscodec.EncodeAccountIDToClassicAddress(channel.DestinationID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode destination address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "PayChannel",
-		"Account":         ownerAddress,
-		"Destination":     destAddress,
-		"Amount":          fmt.Sprintf("%d", channel.Amount),
-		"Balance":         fmt.Sprintf("%d", channel.Balance),
-		"SettleDelay":     channel.SettleDelay,
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode PayChannel: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to parse a PayChannel ledger entry
-func parsePayChannel(data []byte) (*PayChannelData, error) {
-	channel := &PayChannelData{}
-	offset := 0
-
-	for offset < len(data) {
-		if offset+1 > len(data) {
-			break
-		}
-
-		header := data[offset]
-		offset++
-
-		typeCode := (header >> 4) & 0x0F
-		fieldCode := header & 0x0F
-
-		if typeCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			typeCode = data[offset]
-			offset++
-		}
-
-		if fieldCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			fieldCode = data[offset]
-			offset++
-		}
-
-		switch typeCode {
-		case fieldTypeUInt16:
-			if offset+2 > len(data) {
-				return channel, nil
-			}
-			offset += 2
-
-		case fieldTypeUInt32:
-			if offset+4 > len(data) {
-				return channel, nil
-			}
-			value := binary.BigEndian.Uint32(data[offset : offset+4])
-			offset += 4
-			switch fieldCode {
-			case 39: // SettleDelay
-				channel.SettleDelay = value
-			case 37: // CancelAfter
-				channel.CancelAfter = value
-			case 10: // Expiration
-				channel.Expiration = value
-			}
-
-		case fieldTypeUInt64:
-			if offset+8 > len(data) {
-				return channel, nil
-			}
-			offset += 8
-
-		case fieldTypeAmount:
-			if offset+8 > len(data) {
-				return channel, nil
-			}
-			rawAmount := binary.BigEndian.Uint64(data[offset : offset+8])
-			amount := rawAmount & 0x3FFFFFFFFFFFFFFF
-			if fieldCode == 1 { // Amount
-				channel.Amount = amount
-			} else if fieldCode == 5 { // Balance
-				channel.Balance = amount
-			}
-			offset += 8
-
-		case fieldTypeAccountID:
-			if offset+21 > len(data) {
-				return channel, nil
-			}
-			length := data[offset]
-			offset++
-			if length == 20 {
-				switch fieldCode {
-				case 1: // Account
-					copy(channel.Account[:], data[offset:offset+20])
-				case 3: // Destination
-					copy(channel.DestinationID[:], data[offset:offset+20])
-				}
-				offset += 20
-			}
-
-		case fieldTypeBlob:
-			if offset >= len(data) {
-				return channel, nil
-			}
-			length := int(data[offset])
-			offset++
-			if offset+length > len(data) {
-				return channel, nil
-			}
-			if fieldCode == 28 { // PublicKey
-				channel.PublicKey = hex.EncodeToString(data[offset : offset+length])
-			}
-			offset += length
-
-		default:
-			return channel, nil
-		}
-	}
-
-	return channel, nil
-}
-
-// Check ledger entry data
-type CheckData struct {
-	Account       [20]byte
-	DestinationID [20]byte
-	SendMax       uint64 // For XRP checks; IOU checks would need more fields
-	Sequence      uint32
-	Expiration    uint32
-	InvoiceID     [32]byte
-	DestinationTag uint32
-	HasDestTag    bool
-}
-
-// applyCheckCreate applies a CheckCreate transaction
-func (e *Engine) applyCheckCreate(tx *CheckCreate, account *AccountRoot, metadata *Metadata) Result {
-	// Verify destination exists
-	destID, err := decodeAccountID(tx.Destination)
-	if err != nil {
-		return TemINVALID
-	}
-
-	destKey := keylet.Account(destID)
-	exists, _ := e.view.Exists(destKey)
-	if !exists {
-		return TecNO_DST
-	}
-
-	// Parse SendMax - only XRP supported for now
-	sendMax, err := strconv.ParseUint(tx.SendMax.Value, 10, 64)
-	if err != nil {
-		// May be an IOU amount
-		sendMax = 0
-	}
-
-	// Check balance for XRP checks
-	if tx.SendMax.Currency == "" && sendMax > 0 {
-		if account.Balance < sendMax {
-			return TecUNFUNDED
-		}
-	}
-
-	// Create the check entry
-	accountID, _ := decodeAccountID(tx.Account)
-	sequence := *tx.GetCommon().Sequence
-
-	checkKey := keylet.Check(accountID, sequence)
-
-	// Serialize check
-	checkData, err := serializeCheck(tx, accountID, destID, sequence, sendMax)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Insert check
-	if err := e.view.Insert(checkKey, checkData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Increase owner count
-	account.OwnerCount++
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "Check",
-		LedgerIndex:     hex.EncodeToString(checkKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":     tx.Account,
-			"Destination": tx.Destination,
-			"SendMax":     tx.SendMax.Value,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyCheckCash applies a CheckCash transaction
-func (e *Engine) applyCheckCash(tx *CheckCash, account *AccountRoot, metadata *Metadata) Result {
-	// Parse check ID
-	checkID, err := hex.DecodeString(tx.CheckID)
-	if err != nil || len(checkID) != 32 {
-		return TemINVALID
-	}
-
-	var checkKeyBytes [32]byte
-	copy(checkKeyBytes[:], checkID)
-	checkKey := keylet.Keylet{Key: checkKeyBytes}
-
-	// Read check
-	checkData, err := e.view.Read(checkKey)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	// Parse check
-	check, err := parseCheck(checkData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Verify the account is the destination
-	accountID, _ := decodeAccountID(tx.Account)
-	if check.DestinationID != accountID {
-		return TecNO_PERMISSION
-	}
-
-	// Check expiration
-	if check.Expiration > 0 {
-		// In full implementation, check against close time
-		// For standalone mode, we'll allow it
-	}
-
-	// Determine amount to cash
-	var cashAmount uint64
-	if tx.Amount != nil {
-		// Exact amount
-		cashAmount, err = strconv.ParseUint(tx.Amount.Value, 10, 64)
-		if err != nil {
-			return TemINVALID
-		}
-		if cashAmount > check.SendMax {
-			return TecPATH_PARTIAL
-		}
-	} else if tx.DeliverMin != nil {
-		// Minimum amount - use full SendMax for simplicity
-		deliverMin, err := strconv.ParseUint(tx.DeliverMin.Value, 10, 64)
-		if err != nil {
-			return TemINVALID
-		}
-		if check.SendMax < deliverMin {
-			return TecPATH_PARTIAL
-		}
-		cashAmount = check.SendMax
-	}
-
-	// Get the check creator's account
-	creatorKey := keylet.Account(check.Account)
-	creatorData, err := e.view.Read(creatorKey)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	creatorAccount, err := parseAccountRoot(creatorData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Check if creator has sufficient balance
-	if creatorAccount.Balance < cashAmount {
-		return TecUNFUNDED_PAYMENT
-	}
-
-	// Transfer the funds
-	creatorAccount.Balance -= cashAmount
-	account.Balance += cashAmount
-
-	// Decrease creator's owner count
-	if creatorAccount.OwnerCount > 0 {
-		creatorAccount.OwnerCount--
-	}
-
-	// Update creator account
-	creatorUpdatedData, err := serializeAccountRoot(creatorAccount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Update(creatorKey, creatorUpdatedData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Delete the check
-	if err := e.view.Erase(checkKey); err != nil {
-		return TefINTERNAL
-	}
-
-	// Record in metadata
-	creatorAddr, _ := encodeAccountID(check.Account)
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Check",
-		LedgerIndex:     hex.EncodeToString(checkKey.Key[:]),
-	})
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AccountRoot",
-		LedgerIndex:     hex.EncodeToString(creatorKey.Key[:]),
-		FinalFields: map[string]any{
-			"Account": creatorAddr,
-			"Balance": strconv.FormatUint(creatorAccount.Balance, 10),
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyCheckCancel applies a CheckCancel transaction
-func (e *Engine) applyCheckCancel(tx *CheckCancel, account *AccountRoot, metadata *Metadata) Result {
-	// Parse check ID
-	checkID, err := hex.DecodeString(tx.CheckID)
-	if err != nil || len(checkID) != 32 {
-		return TemINVALID
-	}
-
-	var checkKeyBytes [32]byte
-	copy(checkKeyBytes[:], checkID)
-	checkKey := keylet.Keylet{Key: checkKeyBytes}
-
-	// Read check
-	checkData, err := e.view.Read(checkKey)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	// Parse check
-	check, err := parseCheck(checkData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	accountID, _ := decodeAccountID(tx.Account)
-	isCreator := check.Account == accountID
-	isDestination := check.DestinationID == accountID
-
-	// Only creator or destination can cancel
-	if !isCreator && !isDestination {
-		// Unless the check is expired
-		if check.Expiration == 0 {
-			return TecNO_PERMISSION
-		}
-		// In full implementation, check if expired
-		// For standalone mode, allow anyone to cancel expired checks
-	}
-
-	// Delete the check
-	if err := e.view.Erase(checkKey); err != nil {
-		return TefINTERNAL
-	}
-
-	// If the canceller is also the creator, decrease their owner count
-	if isCreator {
-		if account.OwnerCount > 0 {
-			account.OwnerCount--
-		}
-	} else {
-		// Need to update the creator's owner count
-		creatorKey := keylet.Account(check.Account)
-		creatorData, err := e.view.Read(creatorKey)
-		if err == nil {
-			creatorAccount, err := parseAccountRoot(creatorData)
-			if err == nil && creatorAccount.OwnerCount > 0 {
-				creatorAccount.OwnerCount--
-				creatorUpdatedData, _ := serializeAccountRoot(creatorAccount)
-				e.view.Update(creatorKey, creatorUpdatedData)
-			}
-		}
-	}
-
-	// Record in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Check",
-		LedgerIndex:     hex.EncodeToString(checkKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// Helper function to serialize a Check
-func serializeCheck(tx *CheckCreate, ownerID, destID [20]byte, sequence uint32, sendMax uint64) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	destAddress, err := addresscodec.EncodeAccountIDToClassicAddress(destID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode destination address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "Check",
-		"Account":         ownerAddress,
-		"Destination":     destAddress,
-		"SendMax":         fmt.Sprintf("%d", sendMax),
-		"Sequence":        sequence,
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	if tx.Expiration != nil {
-		jsonObj["Expiration"] = *tx.Expiration
-	}
-
-	if tx.DestinationTag != nil {
-		jsonObj["DestinationTag"] = *tx.DestinationTag
-	}
-
-	if tx.InvoiceID != "" {
-		jsonObj["InvoiceID"] = tx.InvoiceID
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode Check: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to parse a Check ledger entry
-func parseCheck(data []byte) (*CheckData, error) {
-	check := &CheckData{}
-	offset := 0
-
-	for offset < len(data) {
-		if offset+1 > len(data) {
-			break
-		}
-
-		header := data[offset]
-		offset++
-
-		typeCode := (header >> 4) & 0x0F
-		fieldCode := header & 0x0F
-
-		if typeCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			typeCode = data[offset]
-			offset++
-		}
-
-		if fieldCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			fieldCode = data[offset]
-			offset++
-		}
-
-		switch typeCode {
-		case fieldTypeUInt16:
-			if offset+2 > len(data) {
-				return check, nil
-			}
-			offset += 2
-
-		case fieldTypeUInt32:
-			if offset+4 > len(data) {
-				return check, nil
-			}
-			value := binary.BigEndian.Uint32(data[offset : offset+4])
-			offset += 4
-			switch fieldCode {
-			case fieldCodeSequence:
-				check.Sequence = value
-			case 10: // Expiration
-				check.Expiration = value
-			case 14: // DestinationTag
-				check.DestinationTag = value
-				check.HasDestTag = true
-			}
-
-		case fieldTypeUInt64:
-			if offset+8 > len(data) {
-				return check, nil
-			}
-			offset += 8
-
-		case fieldTypeHash256:
-			if offset+32 > len(data) {
-				return check, nil
-			}
-			if fieldCode == 17 { // InvoiceID
-				copy(check.InvoiceID[:], data[offset:offset+32])
-			}
-			offset += 32
-
-		case fieldTypeAmount:
-			if offset+8 > len(data) {
-				return check, nil
-			}
-			if data[offset]&0x80 == 0 {
-				// XRP amount
-				rawAmount := binary.BigEndian.Uint64(data[offset : offset+8])
-				if fieldCode == 9 { // SendMax
-					check.SendMax = rawAmount & 0x3FFFFFFFFFFFFFFF
-				}
-				offset += 8
-			} else {
-				// IOU amount - skip 48 bytes
-				offset += 48
-			}
-
-		case fieldTypeAccountID:
-			if offset+21 > len(data) {
-				return check, nil
-			}
-			length := data[offset]
-			offset++
-			if length == 20 {
-				switch fieldCode {
-				case 1: // Account
-					copy(check.Account[:], data[offset:offset+20])
-				case 3: // Destination
-					copy(check.DestinationID[:], data[offset:offset+20])
-				}
-				offset += 20
-			}
-
-		case fieldTypeBlob:
-			if offset >= len(data) {
-				return check, nil
-			}
-			length := int(data[offset])
-			offset++
-			if offset+length > len(data) {
-				return check, nil
-			}
-			offset += length
-
-		default:
-			return check, nil
-		}
-	}
-
-	return check, nil
-}
-
-// NFToken data structures
-
-// NFTokenPageData represents an NFToken page ledger entry
-type NFTokenPageData struct {
-	PreviousPageMin [32]byte
-	NextPageMin     [32]byte
-	NFTokens        []NFTokenData
-}
-
-// NFTokenData represents an individual NFToken within a page
-type NFTokenData struct {
-	NFTokenID [32]byte
-	URI       string
-}
-
-// NFTokenOfferData represents an NFToken offer ledger entry
-type NFTokenOfferData struct {
-	Owner         [20]byte
-	NFTokenID     [32]byte
-	Amount        uint64
-	Flags         uint32
-	Destination   [20]byte
-	Expiration    uint32
-	HasDestination bool
-}
-
-// generateNFTokenID generates an NFTokenID based on the minting parameters
-func generateNFTokenID(issuer [20]byte, taxon, sequence uint32, flags uint16, transferFee uint16) [32]byte {
-	var tokenID [32]byte
-
-	// NFTokenID format (32 bytes):
-	// Bytes 0-1: Flags (2 bytes)
-	// Bytes 2-3: TransferFee (2 bytes)
-	// Bytes 4-23: Issuer AccountID (20 bytes)
-	// Bytes 24-27: Taxon (scrambled, 4 bytes)
-	// Bytes 28-31: Sequence (4 bytes)
-
-	binary.BigEndian.PutUint16(tokenID[0:2], flags)
-	binary.BigEndian.PutUint16(tokenID[2:4], transferFee)
-	copy(tokenID[4:24], issuer[:])
-
-	// Scramble the taxon to prevent enumeration
-	scrambledTaxon := taxon ^ (sequence & 0xFFFFFFFF)
-	binary.BigEndian.PutUint32(tokenID[24:28], scrambledTaxon)
-	binary.BigEndian.PutUint32(tokenID[28:32], sequence)
-
-	return tokenID
-}
-
-// applyNFTokenMint applies an NFTokenMint transaction
-func (e *Engine) applyNFTokenMint(tx *NFTokenMint, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Determine the issuer
-	var issuerID [20]byte
-	if tx.Issuer != "" {
-		var err error
-		issuerID, err = decodeAccountID(tx.Issuer)
-		if err != nil {
-			return TemINVALID
-		}
-	} else {
-		issuerID = accountID
-	}
-
-	// Get flags for the token
-	txFlags := tx.GetFlags()
-	var tokenFlags uint16
-	if txFlags&NFTokenMintFlagBurnable != 0 {
-		tokenFlags |= 0x0001
-	}
-	if txFlags&NFTokenMintFlagOnlyXRP != 0 {
-		tokenFlags |= 0x0002
-	}
-	if txFlags&NFTokenMintFlagTrustLine != 0 {
-		tokenFlags |= 0x0004
-	}
-	if txFlags&NFTokenMintFlagTransferable != 0 {
-		tokenFlags |= 0x0008
-	}
-
-	// Get transfer fee
-	var transferFee uint16
-	if tx.TransferFee != nil {
-		transferFee = *tx.TransferFee
-	}
-
-	// Generate the NFTokenID
-	sequence := *tx.GetCommon().Sequence
-	tokenID := generateNFTokenID(issuerID, tx.NFTokenTaxon, sequence, tokenFlags, transferFee)
-
-	// Find or create the NFToken page for this account
-	// NFToken pages are keyed by account + portion of token ID
-	pageKey := keylet.NFTokenPage(accountID, tokenID)
-
-	// Check if page exists
-	exists, _ := e.view.Exists(pageKey)
-	if exists {
-		// Read existing page and add token
-		pageData, err := e.view.Read(pageKey)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		// Parse the page
-		page, err := parseNFTokenPage(pageData)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		// Add the new token
-		newToken := NFTokenData{
-			NFTokenID: tokenID,
-			URI:       tx.URI,
-		}
-		page.NFTokens = append(page.NFTokens, newToken)
-
-		// Serialize and update
-		updatedPageData, err := serializeNFTokenPage(page)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		if err := e.view.Update(pageKey, updatedPageData); err != nil {
-			return TefINTERNAL
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "NFTokenPage",
-			LedgerIndex:     hex.EncodeToString(pageKey.Key[:]),
-		})
-	} else {
-		// Create new page
-		page := &NFTokenPageData{
-			NFTokens: []NFTokenData{
-				{
-					NFTokenID: tokenID,
-					URI:       tx.URI,
-				},
-			},
-		}
-
-		pageData, err := serializeNFTokenPage(page)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		if err := e.view.Insert(pageKey, pageData); err != nil {
-			return TefINTERNAL
-		}
-
-		// Increase owner count for the new page
-		account.OwnerCount++
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "NFTokenPage",
-			LedgerIndex:     hex.EncodeToString(pageKey.Key[:]),
-			NewFields: map[string]any{
-				"NFTokenID": hex.EncodeToString(tokenID[:]),
-			},
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// applyNFTokenBurn applies an NFTokenBurn transaction
-func (e *Engine) applyNFTokenBurn(tx *NFTokenBurn, account *AccountRoot, metadata *Metadata) Result {
-	// Parse the token ID
-	tokenIDBytes, err := hex.DecodeString(tx.NFTokenID)
-	if err != nil || len(tokenIDBytes) != 32 {
-		return TemINVALID
-	}
-
-	var tokenID [32]byte
-	copy(tokenID[:], tokenIDBytes)
-
-	// Determine the owner
-	var ownerID [20]byte
-	if tx.Owner != "" {
-		ownerID, err = decodeAccountID(tx.Owner)
-		if err != nil {
-			return TemINVALID
-		}
-	} else {
-		ownerID, _ = decodeAccountID(tx.Account)
-	}
-
-	// Find the NFToken page
-	pageKey := keylet.NFTokenPage(ownerID, tokenID)
-
-	pageData, err := e.view.Read(pageKey)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	// Parse the page
-	page, err := parseNFTokenPage(pageData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Find and remove the token
-	found := false
-	for i, token := range page.NFTokens {
-		if token.NFTokenID == tokenID {
-			// Remove token from page
-			page.NFTokens = append(page.NFTokens[:i], page.NFTokens[i+1:]...)
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		return TecNO_ENTRY
-	}
-
-	// Update or delete the page
-	if len(page.NFTokens) == 0 {
-		// Delete empty page
-		if err := e.view.Erase(pageKey); err != nil {
-			return TefINTERNAL
-		}
-
-		if account.OwnerCount > 0 {
-			account.OwnerCount--
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "NFTokenPage",
-			LedgerIndex:     hex.EncodeToString(pageKey.Key[:]),
-		})
-	} else {
-		// Update page
-		updatedPageData, err := serializeNFTokenPage(page)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		if err := e.view.Update(pageKey, updatedPageData); err != nil {
-			return TefINTERNAL
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "NFTokenPage",
-			LedgerIndex:     hex.EncodeToString(pageKey.Key[:]),
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// applyNFTokenCreateOffer applies an NFTokenCreateOffer transaction
-func (e *Engine) applyNFTokenCreateOffer(tx *NFTokenCreateOffer, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Parse token ID
-	tokenIDBytes, err := hex.DecodeString(tx.NFTokenID)
-	if err != nil || len(tokenIDBytes) != 32 {
-		return TemINVALID
-	}
-
-	var tokenID [32]byte
-	copy(tokenID[:], tokenIDBytes)
-
-	// Parse amount (XRP only for now)
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err != nil {
-		amount = 0
-	}
-
-	// Check if this is a sell offer
-	isSellOffer := tx.GetFlags()&NFTokenCreateOfferFlagSellNFToken != 0
-
-	if isSellOffer {
-		// For sell offers, verify the sender owns the token
-		pageKey := keylet.NFTokenPage(accountID, tokenID)
-		_, err := e.view.Read(pageKey)
-		if err != nil {
-			return TecNO_ENTRY
-		}
-	} else {
-		// For buy offers, need to escrow the funds (XRP)
-		if tx.Amount.Currency == "" && amount > 0 {
-			if account.Balance < amount {
-				return TecUNFUNDED
-			}
-			account.Balance -= amount
-		}
-	}
-
-	// Create the offer
-	sequence := *tx.GetCommon().Sequence
-	offerKey := keylet.NFTokenOffer(accountID, sequence)
-
-	offerData, err := serializeNFTokenOffer(tx, accountID, tokenID, amount, sequence)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Insert(offerKey, offerData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Increase owner count
-	account.OwnerCount++
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "NFTokenOffer",
-		LedgerIndex:     hex.EncodeToString(offerKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":   tx.Account,
-			"NFTokenID": tx.NFTokenID,
-			"Amount":    tx.Amount.Value,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyNFTokenCancelOffer applies an NFTokenCancelOffer transaction
-func (e *Engine) applyNFTokenCancelOffer(tx *NFTokenCancelOffer, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	for _, offerIDHex := range tx.NFTokenOffers {
-		// Parse offer ID
-		offerIDBytes, err := hex.DecodeString(offerIDHex)
-		if err != nil || len(offerIDBytes) != 32 {
-			continue
-		}
-
-		var offerKeyBytes [32]byte
-		copy(offerKeyBytes[:], offerIDBytes)
-		offerKey := keylet.Keylet{Key: offerKeyBytes}
-
-		// Read the offer
-		offerData, err := e.view.Read(offerKey)
-		if err != nil {
-			continue // Skip non-existent offers
-		}
-
-		// Parse the offer
-		offer, err := parseNFTokenOffer(offerData)
-		if err != nil {
-			continue
-		}
-
-		// Verify the canceller is the owner or the offer expired
-		if offer.Owner != accountID {
-			// Check if offer has expired (in full implementation)
-			// For standalone, allow owner to cancel
-			continue
-		}
-
-		// If this was a buy offer, refund the escrowed amount
-		if offer.Flags&uint32(NFTokenCreateOfferFlagSellNFToken) == 0 {
-			// Buy offer - refund
-			account.Balance += offer.Amount
-		}
-
-		// Delete the offer
-		if err := e.view.Erase(offerKey); err != nil {
-			continue
-		}
-
-		if account.OwnerCount > 0 {
-			account.OwnerCount--
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "NFTokenOffer",
-			LedgerIndex:     hex.EncodeToString(offerKey.Key[:]),
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// applyNFTokenAcceptOffer applies an NFTokenAcceptOffer transaction
-func (e *Engine) applyNFTokenAcceptOffer(tx *NFTokenAcceptOffer, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Handle sell offer acceptance
-	if tx.NFTokenSellOffer != "" && tx.NFTokenBuyOffer == "" {
-		// Accept a sell offer
-		sellOfferIDBytes, err := hex.DecodeString(tx.NFTokenSellOffer)
-		if err != nil || len(sellOfferIDBytes) != 32 {
-			return TemINVALID
-		}
-
-		var sellOfferKeyBytes [32]byte
-		copy(sellOfferKeyBytes[:], sellOfferIDBytes)
-		sellOfferKey := keylet.Keylet{Key: sellOfferKeyBytes}
-
-		// Read sell offer
-		sellOfferData, err := e.view.Read(sellOfferKey)
-		if err != nil {
-			return TecOBJECT_NOT_FOUND
-		}
-
-		sellOffer, err := parseNFTokenOffer(sellOfferData)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		// Check if destination matches (if set)
-		if sellOffer.HasDestination && sellOffer.Destination != accountID {
-			return TecNO_PERMISSION
-		}
-
-		// Pay for the NFT
-		if sellOffer.Amount > 0 {
-			if account.Balance < sellOffer.Amount {
-				return TecUNFUNDED_PAYMENT
-			}
-			account.Balance -= sellOffer.Amount
-
-			// Pay the seller
-			sellerKey := keylet.Account(sellOffer.Owner)
-			sellerData, err := e.view.Read(sellerKey)
-			if err != nil {
-				return TefINTERNAL
-			}
-
-			sellerAccount, err := parseAccountRoot(sellerData)
-			if err != nil {
-				return TefINTERNAL
-			}
-
-			sellerAccount.Balance += sellOffer.Amount
-			if sellerAccount.OwnerCount > 0 {
-				sellerAccount.OwnerCount-- // For the offer being deleted
-			}
-
-			sellerUpdatedData, err := serializeAccountRoot(sellerAccount)
-			if err != nil {
-				return TefINTERNAL
-			}
-
-			if err := e.view.Update(sellerKey, sellerUpdatedData); err != nil {
-				return TefINTERNAL
-			}
-		}
-
-		// Transfer the NFT (simplified - just update metadata)
-		// In full implementation, would move token between pages
-
-		// Delete the sell offer
-		if err := e.view.Erase(sellOfferKey); err != nil {
-			return TefINTERNAL
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "NFTokenOffer",
-			LedgerIndex:     hex.EncodeToString(sellOfferKey.Key[:]),
-		})
-
-		return TesSUCCESS
-	}
-
-	// Handle buy offer acceptance
-	if tx.NFTokenBuyOffer != "" && tx.NFTokenSellOffer == "" {
-		// Accept a buy offer
-		buyOfferIDBytes, err := hex.DecodeString(tx.NFTokenBuyOffer)
-		if err != nil || len(buyOfferIDBytes) != 32 {
-			return TemINVALID
-		}
-
-		var buyOfferKeyBytes [32]byte
-		copy(buyOfferKeyBytes[:], buyOfferIDBytes)
-		buyOfferKey := keylet.Keylet{Key: buyOfferKeyBytes}
-
-		// Read buy offer
-		buyOfferData, err := e.view.Read(buyOfferKey)
-		if err != nil {
-			return TecOBJECT_NOT_FOUND
-		}
-
-		buyOffer, err := parseNFTokenOffer(buyOfferData)
-		if err != nil {
-			return TefINTERNAL
-		}
-
-		// Receive payment (already escrowed in buy offer)
-		account.Balance += buyOffer.Amount
-
-		// Decrease buyer's owner count
-		buyerKey := keylet.Account(buyOffer.Owner)
-		buyerData, err := e.view.Read(buyerKey)
-		if err == nil {
-			buyerAccount, err := parseAccountRoot(buyerData)
-			if err == nil && buyerAccount.OwnerCount > 0 {
-				buyerAccount.OwnerCount--
-				buyerUpdatedData, _ := serializeAccountRoot(buyerAccount)
-				e.view.Update(buyerKey, buyerUpdatedData)
-			}
-		}
-
-		// Transfer the NFT (simplified)
-		// In full implementation, would move token between pages
-
-		// Delete the buy offer
-		if err := e.view.Erase(buyOfferKey); err != nil {
-			return TefINTERNAL
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "NFTokenOffer",
-			LedgerIndex:     hex.EncodeToString(buyOfferKey.Key[:]),
-		})
-
-		return TesSUCCESS
-	}
-
-	// Brokered mode (both offers) - simplified implementation
-	if tx.NFTokenSellOffer != "" && tx.NFTokenBuyOffer != "" {
-		// This would involve matching a buy and sell offer
-		// Simplified: just delete both offers and transfer funds
-		return TesSUCCESS
-	}
-
-	return TemINVALID
-}
-
-// Helper function to serialize an NFToken page
-func serializeNFTokenPage(page *NFTokenPageData) ([]byte, error) {
-	jsonObj := map[string]any{
-		"LedgerEntryType": "NFTokenPage",
-		"Flags":           uint32(0),
-	}
-
-	var emptyHash [32]byte
-	if page.PreviousPageMin != emptyHash {
-		jsonObj["PreviousPageMin"] = strings.ToUpper(hex.EncodeToString(page.PreviousPageMin[:]))
-	}
-
-	if page.NextPageMin != emptyHash {
-		jsonObj["NextPageMin"] = strings.ToUpper(hex.EncodeToString(page.NextPageMin[:]))
-	}
-
-	// Build NFTokens array
-	if len(page.NFTokens) > 0 {
-		nfTokens := make([]map[string]any, len(page.NFTokens))
-		for i, token := range page.NFTokens {
-			nfToken := map[string]any{
-				"NFToken": map[string]any{
-					"NFTokenID": strings.ToUpper(hex.EncodeToString(token.NFTokenID[:])),
-				},
-			}
-			if token.URI != "" {
-				nfToken["NFToken"].(map[string]any)["URI"] = token.URI
-			}
-			nfTokens[i] = nfToken
-		}
-		jsonObj["NFTokens"] = nfTokens
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode NFTokenPage: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to parse an NFToken page
-func parseNFTokenPage(data []byte) (*NFTokenPageData, error) {
-	page := &NFTokenPageData{
-		NFTokens: make([]NFTokenData, 0),
-	}
-	offset := 0
-
-	var currentToken NFTokenData
-	hasToken := false
-
-	for offset < len(data) {
-		if offset+1 > len(data) {
-			break
-		}
-
-		header := data[offset]
-		offset++
-
-		typeCode := (header >> 4) & 0x0F
-		fieldCode := header & 0x0F
-
-		if typeCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			typeCode = data[offset]
-			offset++
-		}
-
-		if fieldCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			fieldCode = data[offset]
-			offset++
-		}
-
-		switch typeCode {
-		case fieldTypeUInt16:
-			if offset+2 > len(data) {
-				return page, nil
-			}
-			offset += 2
-
-		case fieldTypeUInt32:
-			if offset+4 > len(data) {
-				return page, nil
-			}
-			offset += 4
-
-		case fieldTypeHash256:
-			if offset+32 > len(data) {
-				return page, nil
-			}
-			switch fieldCode {
-			case 25: // PreviousPageMin
-				copy(page.PreviousPageMin[:], data[offset:offset+32])
-			case 26: // NextPageMin
-				copy(page.NextPageMin[:], data[offset:offset+32])
-			case 10: // NFTokenID
-				if hasToken {
-					page.NFTokens = append(page.NFTokens, currentToken)
-				}
-				copy(currentToken.NFTokenID[:], data[offset:offset+32])
-				currentToken.URI = ""
-				hasToken = true
-			}
-			offset += 32
-
-		case fieldTypeBlob:
-			if offset >= len(data) {
-				return page, nil
-			}
-			length := int(data[offset])
-			offset++
-			if offset+length > len(data) {
-				return page, nil
-			}
-			if fieldCode == 5 { // URI
-				currentToken.URI = hex.EncodeToString(data[offset : offset+length])
-			}
-			offset += length
-
-		default:
-			if hasToken {
-				page.NFTokens = append(page.NFTokens, currentToken)
-				hasToken = false
-			}
-			return page, nil
-		}
-	}
-
-	if hasToken {
-		page.NFTokens = append(page.NFTokens, currentToken)
-	}
-
-	return page, nil
-}
-
-// Helper function to serialize an NFToken offer
-func serializeNFTokenOffer(tx *NFTokenCreateOffer, ownerID [20]byte, tokenID [32]byte, amount uint64, sequence uint32) ([]byte, error) {
-	ownerAddress, err := addresscodec.EncodeAccountIDToClassicAddress(ownerID[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode owner address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "NFTokenOffer",
-		"Account":         ownerAddress,
-		"Amount":          fmt.Sprintf("%d", amount),
-		"NFTokenID":       strings.ToUpper(hex.EncodeToString(tokenID[:])),
-		"OwnerNode":       "0",
-		"Flags":           tx.GetFlags(),
-	}
-
-	if tx.Expiration != nil {
-		jsonObj["Expiration"] = *tx.Expiration
-	}
-
-	if tx.Destination != "" {
-		jsonObj["Destination"] = tx.Destination
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode NFTokenOffer: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Helper function to parse an NFToken offer
-func parseNFTokenOffer(data []byte) (*NFTokenOfferData, error) {
-	offer := &NFTokenOfferData{}
-	offset := 0
-
-	for offset < len(data) {
-		if offset+1 > len(data) {
-			break
-		}
-
-		header := data[offset]
-		offset++
-
-		typeCode := (header >> 4) & 0x0F
-		fieldCode := header & 0x0F
-
-		if typeCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			typeCode = data[offset]
-			offset++
-		}
-
-		if fieldCode == 0 {
-			if offset >= len(data) {
-				break
-			}
-			fieldCode = data[offset]
-			offset++
-		}
-
-		switch typeCode {
-		case fieldTypeUInt16:
-			if offset+2 > len(data) {
-				return offer, nil
-			}
-			offset += 2
-
-		case fieldTypeUInt32:
-			if offset+4 > len(data) {
-				return offer, nil
-			}
-			value := binary.BigEndian.Uint32(data[offset : offset+4])
-			offset += 4
-			switch fieldCode {
-			case fieldCodeFlags:
-				offer.Flags = value
-			case 10: // Expiration
-				offer.Expiration = value
-			}
-
-		case fieldTypeUInt64:
-			if offset+8 > len(data) {
-				return offer, nil
-			}
-			offset += 8
-
-		case fieldTypeHash256:
-			if offset+32 > len(data) {
-				return offer, nil
-			}
-			if fieldCode == 10 { // NFTokenID
-				copy(offer.NFTokenID[:], data[offset:offset+32])
-			}
-			offset += 32
-
-		case fieldTypeAmount:
-			if offset+8 > len(data) {
-				return offer, nil
-			}
-			if data[offset]&0x80 == 0 {
-				rawAmount := binary.BigEndian.Uint64(data[offset : offset+8])
-				offer.Amount = rawAmount & 0x3FFFFFFFFFFFFFFF
-				offset += 8
-			} else {
-				offset += 48
-			}
-
-		case fieldTypeAccountID:
-			if offset+21 > len(data) {
-				return offer, nil
-			}
-			length := data[offset]
-			offset++
-			if length == 20 {
-				switch fieldCode {
-				case 1: // Account/Owner
-					copy(offer.Owner[:], data[offset:offset+20])
-				case 3: // Destination
-					copy(offer.Destination[:], data[offset:offset+20])
-					offer.HasDestination = true
-				}
-				offset += 20
-			}
-
-		default:
-			return offer, nil
-		}
-	}
-
-	return offer, nil
-}
-
-// AMM data structures
-
-// AMMData represents an AMM ledger entry
-type AMMData struct {
-	Account       [20]byte // AMM account
-	Asset         [20]byte // First asset currency (20 bytes)
-	Asset2        [20]byte // Second asset currency (20 bytes)
-	TradingFee    uint16
-	LPTokenBalance uint64
-	VoteSlots     []VoteSlotData
-	AuctionSlot   *AuctionSlotData
-}
-
-// VoteSlotData represents a voting slot in an AMM
-type VoteSlotData struct {
-	Account    [20]byte
-	TradingFee uint16
-	VoteWeight uint32
-}
-
-// AuctionSlotData represents the auction slot in an AMM
-type AuctionSlotData struct {
-	Account       [20]byte
-	Expiration    uint32
-	Price         uint64
-	AuthAccounts  [][20]byte
-}
-
-// computeAMMAccountID derives the AMM account from the asset pair
-func computeAMMAccountID(asset1, asset2 Asset) [20]byte {
-	// In rippled, this is computed by hashing the asset pair
-	// Simplified implementation: hash the currency codes
-	var result [20]byte
-	data := []byte(asset1.Currency + asset1.Issuer + asset2.Currency + asset2.Issuer)
-	hash := crypto.Sha512Half(data)
-	copy(result[:], hash[:20])
-	return result
-}
-
-// applyAMMCreate applies an AMMCreate transaction
-func (e *Engine) applyAMMCreate(tx *AMMCreate, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Parse amounts
-	amount1, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err != nil {
-		amount1 = 0
-	}
-	amount2, err := strconv.ParseUint(tx.Amount2.Value, 10, 64)
-	if err != nil {
-		amount2 = 0
-	}
-
-	// For XRP amounts, check balance
-	if tx.Amount.Currency == "" && amount1 > 0 {
-		if account.Balance < amount1 {
-			return TecUNFUNDED
-		}
-		account.Balance -= amount1
-	}
-	if tx.Amount2.Currency == "" && amount2 > 0 {
-		if account.Balance < amount2 {
-			return TecUNFUNDED
-		}
-		account.Balance -= amount2
-	}
-
-	// Compute AMM account ID
-	asset1 := Asset{Currency: tx.Amount.Currency, Issuer: tx.Amount.Issuer}
-	asset2 := Asset{Currency: tx.Amount2.Currency, Issuer: tx.Amount2.Issuer}
-	ammAccountID := computeAMMAccountID(asset1, asset2)
-
-	// Check if AMM already exists
-	ammKey := keylet.Account(ammAccountID)
-	exists, _ := e.view.Exists(ammKey)
-	if exists {
-		return TecDUPLICATE
-	}
-
-	// Create the AMM account
-	ammAccount := &AccountRoot{
-		Balance:  0,
-		Sequence: 0,
-		Flags:    0,
-	}
-	ammAccount.Account, _ = encodeAccountID(ammAccountID)
-
-	ammAccountData, err := serializeAccountRoot(ammAccount)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	if err := e.view.Insert(ammKey, ammAccountData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Create the AMM entry (simplified)
-	ammData := &AMMData{
-		Account:        ammAccountID,
-		TradingFee:     tx.TradingFee,
-		LPTokenBalance: amount1 + amount2, // Simplified LP calculation
-	}
-
-	ammEntryData, err := serializeAMM(ammData, accountID)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Use AMM keylet (simplified - using account keylet)
-	if err := e.view.Insert(ammKey, ammEntryData); err != nil {
-		return TefINTERNAL
-	}
-
-	// Increase owner count (for AMM and LP tokens)
-	account.OwnerCount++
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":    ammAccount.Account,
-			"TradingFee": tx.TradingFee,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyAMMDeposit applies an AMMDeposit transaction
-func (e *Engine) applyAMMDeposit(tx *AMMDeposit, account *AccountRoot, metadata *Metadata) Result {
-	// Find the AMM
-	ammAccountID := computeAMMAccountID(tx.Asset, tx.Asset2)
-	ammKey := keylet.Account(ammAccountID)
-
-	exists, _ := e.view.Exists(ammKey)
-	if !exists {
-		return TecNO_ENTRY
-	}
-
-	// Parse deposit amounts
-	if tx.Amount != nil {
-		amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-		if err == nil && tx.Amount.Currency == "" {
-			if account.Balance < amount {
-				return TecUNFUNDED
-			}
-			account.Balance -= amount
-		}
-	}
-	if tx.Amount2 != nil {
-		amount2, err := strconv.ParseUint(tx.Amount2.Value, 10, 64)
-		if err == nil && tx.Amount2.Currency == "" {
-			if account.Balance < amount2 {
-				return TecUNFUNDED
-			}
-			account.Balance -= amount2
-		}
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyAMMWithdraw applies an AMMWithdraw transaction
-func (e *Engine) applyAMMWithdraw(tx *AMMWithdraw, account *AccountRoot, metadata *Metadata) Result {
-	// Find the AMM
-	ammAccountID := computeAMMAccountID(tx.Asset, tx.Asset2)
-	ammKey := keylet.Account(ammAccountID)
-
-	exists, _ := e.view.Exists(ammKey)
-	if !exists {
-		return TecNO_ENTRY
-	}
-
-	// Process withdrawal - simplified, just update metadata
-	if tx.Amount != nil {
-		amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-		if err == nil && tx.Amount.Currency == "" {
-			account.Balance += amount
-		}
-	}
-	if tx.Amount2 != nil {
-		amount2, err := strconv.ParseUint(tx.Amount2.Value, 10, 64)
-		if err == nil && tx.Amount2.Currency == "" {
-			account.Balance += amount2
-		}
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyAMMVote applies an AMMVote transaction
-func (e *Engine) applyAMMVote(tx *AMMVote, account *AccountRoot, metadata *Metadata) Result {
-	// Find the AMM
-	ammAccountID := computeAMMAccountID(tx.Asset, tx.Asset2)
-	ammKey := keylet.Account(ammAccountID)
-
-	exists, _ := e.view.Exists(ammKey)
-	if !exists {
-		return TecNO_ENTRY
-	}
-
-	// Record vote - simplified, in full implementation would update vote slots
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-		FinalFields: map[string]any{
-			"TradingFee": tx.TradingFee,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyAMMBid applies an AMMBid transaction
-func (e *Engine) applyAMMBid(tx *AMMBid, account *AccountRoot, metadata *Metadata) Result {
-	// Find the AMM
-	ammAccountID := computeAMMAccountID(tx.Asset, tx.Asset2)
-	ammKey := keylet.Account(ammAccountID)
-
-	exists, _ := e.view.Exists(ammKey)
-	if !exists {
-		return TecNO_ENTRY
-	}
-
-	// Process bid - simplified
-	if tx.BidMin != nil {
-		bidAmount, err := strconv.ParseUint(tx.BidMin.Value, 10, 64)
-		if err == nil && tx.BidMin.Currency == "" {
-			if account.Balance < bidAmount {
-				return TecUNFUNDED
-			}
-			account.Balance -= bidAmount
-		}
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyAMMDelete applies an AMMDelete transaction
-func (e *Engine) applyAMMDelete(tx *AMMDelete, account *AccountRoot, metadata *Metadata) Result {
-	// Find the AMM
-	ammAccountID := computeAMMAccountID(tx.Asset, tx.Asset2)
-	ammKey := keylet.Account(ammAccountID)
-
-	exists, _ := e.view.Exists(ammKey)
-	if !exists {
-		return TecNO_ENTRY
-	}
-
-	// Delete the AMM (only works if empty)
-	if err := e.view.Erase(ammKey); err != nil {
-		return TefINTERNAL
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyAMMClawback applies an AMMClawback transaction
-func (e *Engine) applyAMMClawback(tx *AMMClawback, account *AccountRoot, metadata *Metadata) Result {
-	// Find the AMM
-	ammAccountID := computeAMMAccountID(tx.Asset, tx.Asset2)
-	ammKey := keylet.Account(ammAccountID)
-
-	exists, _ := e.view.Exists(ammKey)
-	if !exists {
-		return TecNO_ENTRY
-	}
-
-	// Find the holder
-	holderID, err := decodeAccountID(tx.Holder)
-	if err != nil {
-		return TemINVALID
-	}
-
-	holderKey := keylet.Account(holderID)
-	exists, _ = e.view.Exists(holderKey)
-	if !exists {
-		return TecNO_TARGET
-	}
-
-	// Clawback LP tokens - simplified
-	if tx.Amount != nil {
-		clawbackAmount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-		if err == nil {
-			// Transfer clawed back value to issuer
-			account.Balance += clawbackAmount
-		}
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "AMM",
-		LedgerIndex:     hex.EncodeToString(ammKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// Helper function to serialize an AMM entry
-func serializeAMM(amm *AMMData, ownerID [20]byte) ([]byte, error) {
-	accountAddress, err := addresscodec.EncodeAccountIDToClassicAddress(amm.Account[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode AMM account address: %w", err)
-	}
-
-	jsonObj := map[string]any{
-		"LedgerEntryType": "AMM",
-		"Account":         accountAddress,
-		"TradingFee":      amm.TradingFee,
-		"OwnerNode":       "0",
-		"Flags":           uint32(0),
-	}
-
-	hexStr, err := binarycodec.Encode(jsonObj)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode AMM: %w", err)
-	}
-
-	return hex.DecodeString(hexStr)
-}
-
-// Phase 5: XChain, DID, Oracle, MPToken implementations
-
-// applyXChainCreateBridge applies an XChainCreateBridge transaction
-func (e *Engine) applyXChainCreateBridge(tx *XChainCreateBridge, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Create Bridge entry (simplified - in full implementation would create Bridge ledger entry)
-	bridgeKey := keylet.Account(accountID) // Simplified - would use Bridge keylet
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "Bridge",
-		LedgerIndex:     hex.EncodeToString(bridgeKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":         tx.Account,
-			"SignatureReward": tx.SignatureReward.Value,
-		},
-	})
-
-	account.OwnerCount++
-	return TesSUCCESS
-}
-
-// applyXChainModifyBridge applies an XChainModifyBridge transaction
-func (e *Engine) applyXChainModifyBridge(tx *XChainModifyBridge, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	bridgeKey := keylet.Account(accountID)
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "Bridge",
-		LedgerIndex:     hex.EncodeToString(bridgeKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyXChainCreateClaimID applies an XChainCreateClaimID transaction
-func (e *Engine) applyXChainCreateClaimID(tx *XChainCreateClaimID, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	sequence := *tx.GetCommon().Sequence
-
-	// Create XChainClaimID entry
-	claimKey := keylet.Escrow(accountID, sequence) // Simplified
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "XChainOwnedClaimID",
-		LedgerIndex:     hex.EncodeToString(claimKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":          tx.Account,
-			"OtherChainSource": tx.OtherChainSource,
-		},
-	})
-
-	account.OwnerCount++
-	return TesSUCCESS
-}
-
-// applyXChainCommit applies an XChainCommit transaction
-func (e *Engine) applyXChainCommit(tx *XChainCommit, account *AccountRoot, metadata *Metadata) Result {
-	// Lock the amount
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err == nil && tx.Amount.Currency == "" {
-		if account.Balance < amount {
-			return TecUNFUNDED
-		}
-		account.Balance -= amount
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "XChainOwnedClaimID",
-	})
-
-	return TesSUCCESS
-}
-
-// applyXChainClaim applies an XChainClaim transaction
-func (e *Engine) applyXChainClaim(tx *XChainClaim, account *AccountRoot, metadata *Metadata) Result {
-	// Credit the claimed amount
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err == nil && tx.Amount.Currency == "" {
-		// Find destination and credit
-		destID, err := decodeAccountID(tx.Destination)
-		if err != nil {
-			return TemINVALID
-		}
-
-		destKey := keylet.Account(destID)
-		destData, err := e.view.Read(destKey)
-		if err == nil {
-			destAccount, err := parseAccountRoot(destData)
-			if err == nil {
-				destAccount.Balance += amount
-				destUpdatedData, _ := serializeAccountRoot(destAccount)
-				e.view.Update(destKey, destUpdatedData)
-			}
-		}
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "XChainOwnedClaimID",
-	})
-
-	return TesSUCCESS
-}
-
-// applyXChainAccountCreateCommit applies an XChainAccountCreateCommit transaction
-func (e *Engine) applyXChainAccountCreateCommit(tx *XChainAccountCreateCommit, account *AccountRoot, metadata *Metadata) Result {
-	// Lock the amount
-	amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-	if err == nil && tx.Amount.Currency == "" {
-		if account.Balance < amount {
-			return TecUNFUNDED
-		}
-		account.Balance -= amount
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "XChainOwnedCreateAccountClaimID",
-		NewFields: map[string]any{
-			"Account":     tx.Account,
-			"Destination": tx.Destination,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyXChainAddClaimAttestation applies an XChainAddClaimAttestation transaction
-func (e *Engine) applyXChainAddClaimAttestation(tx *XChainAddClaimAttestation, account *AccountRoot, metadata *Metadata) Result {
-	// Add attestation to the claim
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "XChainOwnedClaimID",
-	})
-
-	return TesSUCCESS
-}
-
-// applyXChainAddAccountCreateAttestation applies an XChainAddAccountCreateAttestation transaction
-func (e *Engine) applyXChainAddAccountCreateAttestation(tx *XChainAddAccountCreateAttestation, account *AccountRoot, metadata *Metadata) Result {
-	// Add attestation to the account create claim
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "XChainOwnedCreateAccountClaimID",
-	})
-
-	return TesSUCCESS
-}
-
-// applyDIDSet applies a DIDSet transaction
-func (e *Engine) applyDIDSet(tx *DIDSet, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	didKey := keylet.Account(accountID) // Simplified - would use DID keylet
-
-	exists, _ := e.view.Exists(didKey)
-	if exists {
-		// Update existing DID
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "DID",
-			LedgerIndex:     hex.EncodeToString(didKey.Key[:]),
-			FinalFields: map[string]any{
-				"Account": tx.Account,
-			},
-		})
-	} else {
-		// Create new DID
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "DID",
-			LedgerIndex:     hex.EncodeToString(didKey.Key[:]),
-			NewFields: map[string]any{
-				"Account": tx.Account,
-			},
-		})
-		account.OwnerCount++
-	}
-
-	return TesSUCCESS
-}
-
-// applyDIDDelete applies a DIDDelete transaction
-func (e *Engine) applyDIDDelete(tx *DIDDelete, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	didKey := keylet.Account(accountID)
-
-	if account.OwnerCount > 0 {
-		account.OwnerCount--
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "DID",
-		LedgerIndex:     hex.EncodeToString(didKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyOracleSet applies an OracleSet transaction
-func (e *Engine) applyOracleSet(tx *OracleSet, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	oracleKey := keylet.Escrow(accountID, tx.OracleDocumentID) // Simplified
-
-	exists, _ := e.view.Exists(oracleKey)
-	if exists {
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "Oracle",
-			LedgerIndex:     hex.EncodeToString(oracleKey.Key[:]),
-			FinalFields: map[string]any{
-				"LastUpdateTime": tx.LastUpdateTime,
-			},
-		})
-	} else {
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "Oracle",
-			LedgerIndex:     hex.EncodeToString(oracleKey.Key[:]),
-			NewFields: map[string]any{
-				"Account":          tx.Account,
-				"OracleDocumentID": tx.OracleDocumentID,
-				"Provider":         tx.Provider,
-			},
-		})
-		account.OwnerCount++
-	}
-
-	return TesSUCCESS
-}
-
-// applyOracleDelete applies an OracleDelete transaction
-func (e *Engine) applyOracleDelete(tx *OracleDelete, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	oracleKey := keylet.Escrow(accountID, tx.OracleDocumentID)
-
-	if account.OwnerCount > 0 {
-		account.OwnerCount--
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Oracle",
-		LedgerIndex:     hex.EncodeToString(oracleKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyMPTokenIssuanceCreate applies an MPTokenIssuanceCreate transaction
-func (e *Engine) applyMPTokenIssuanceCreate(tx *MPTokenIssuanceCreate, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-	sequence := *tx.GetCommon().Sequence
-	issuanceKey := keylet.Escrow(accountID, sequence) // Simplified
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "MPTokenIssuance",
-		LedgerIndex:     hex.EncodeToString(issuanceKey.Key[:]),
-		NewFields: map[string]any{
-			"Account":  tx.Account,
-			"Sequence": sequence,
-		},
-	})
-
-	account.OwnerCount++
-	return TesSUCCESS
-}
-
-// applyMPTokenIssuanceDestroy applies an MPTokenIssuanceDestroy transaction
-func (e *Engine) applyMPTokenIssuanceDestroy(tx *MPTokenIssuanceDestroy, account *AccountRoot, metadata *Metadata) Result {
-	// Parse issuance ID
-	issuanceIDBytes, err := hex.DecodeString(tx.MPTokenIssuanceID)
-	if err != nil || len(issuanceIDBytes) != 32 {
-		return TemINVALID
-	}
-
-	var issuanceKeyBytes [32]byte
-	copy(issuanceKeyBytes[:], issuanceIDBytes)
-	issuanceKey := keylet.Keylet{Key: issuanceKeyBytes}
-
-	if account.OwnerCount > 0 {
-		account.OwnerCount--
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "MPTokenIssuance",
-		LedgerIndex:     hex.EncodeToString(issuanceKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyMPTokenIssuanceSet applies an MPTokenIssuanceSet transaction
-func (e *Engine) applyMPTokenIssuanceSet(tx *MPTokenIssuanceSet, account *AccountRoot, metadata *Metadata) Result {
-	// Parse issuance ID
-	issuanceIDBytes, err := hex.DecodeString(tx.MPTokenIssuanceID)
-	if err != nil || len(issuanceIDBytes) != 32 {
-		return TemINVALID
-	}
-
-	var issuanceKeyBytes [32]byte
-	copy(issuanceKeyBytes[:], issuanceIDBytes)
-	issuanceKey := keylet.Keylet{Key: issuanceKeyBytes}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "MPTokenIssuance",
-		LedgerIndex:     hex.EncodeToString(issuanceKey.Key[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyMPTokenAuthorize applies an MPTokenAuthorize transaction
-func (e *Engine) applyMPTokenAuthorize(tx *MPTokenAuthorize, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Parse issuance ID
-	issuanceIDBytes, err := hex.DecodeString(tx.MPTokenIssuanceID)
-	if err != nil || len(issuanceIDBytes) != 32 {
-		return TemINVALID
-	}
-
-	// Create or modify MPToken entry
-	tokenKey := keylet.Account(accountID) // Simplified
-
-	flags := tx.GetFlags()
-	if flags&MPTokenAuthorizeFlagUnauthorize != 0 {
-		// Unauthorized - delete MPToken
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "DeletedNode",
-			LedgerEntryType: "MPToken",
-			LedgerIndex:     hex.EncodeToString(tokenKey.Key[:]),
-		})
-		if account.OwnerCount > 0 {
-			account.OwnerCount--
-		}
-	} else {
-		// Authorize - create or modify MPToken
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "MPToken",
-			LedgerIndex:     hex.EncodeToString(tokenKey.Key[:]),
-			NewFields: map[string]any{
-				"Account":           tx.Account,
-				"MPTokenIssuanceID": tx.MPTokenIssuanceID,
-			},
-		})
-		account.OwnerCount++
-	}
-
-	return TesSUCCESS
-}
-
-// applyClawback applies a Clawback transaction
-func (e *Engine) applyClawback(tx *Clawback, account *AccountRoot, metadata *Metadata) Result {
-	// Parse the amount to claw back
-	if tx.Amount.Value == "" {
-		return TemINVALID
-	}
-
-	// For clawback, we need to find the trust line and adjust the balance
-	// The issuer is clawing back from a holder
-	holderID, err := decodeAccountID(tx.Amount.Issuer)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	issuerID, _ := decodeAccountID(tx.Account)
-
-	// Find the trust line
-	trustKey := keylet.Line(holderID, issuerID, tx.Amount.Currency)
-
-	trustData, err := e.view.Read(trustKey)
-	if err != nil {
-		return TecNO_LINE
-	}
-
-	// Parse and modify the trust line
-	rs, err := parseRippleState(trustData)
-	if err != nil {
-		return TefINTERNAL
-	}
-
-	// Record the clawback in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "RippleState",
-		LedgerIndex:     hex.EncodeToString(trustKey.Key[:]),
-		FinalFields: map[string]any{
-			"Balance": rs.Balance,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyNFTokenModify applies an NFTokenModify transaction
-func (e *Engine) applyNFTokenModify(tx *NFTokenModify, account *AccountRoot, metadata *Metadata) Result {
-	if tx.NFTokenID == "" {
-		return TemINVALID
-	}
-
-	// Find the NFToken and modify its URI
-	tokenIDBytes, err := hex.DecodeString(tx.NFTokenID)
-	if err != nil || len(tokenIDBytes) != 32 {
-		return TemINVALID
-	}
-
-	// Record the modification in metadata
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "NFTokenPage",
-		FinalFields: map[string]any{
-			"NFTokenID": tx.NFTokenID,
-			"URI":       tx.URI,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyCredentialCreate applies a CredentialCreate transaction
-func (e *Engine) applyCredentialCreate(tx *CredentialCreate, account *AccountRoot, metadata *Metadata) Result {
-	if tx.Subject == "" || tx.CredentialType == "" {
-		return TemINVALID
-	}
-
-	subjectID, err := decodeAccountID(tx.Subject)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	issuerID, _ := decodeAccountID(tx.Account)
-
-	// Create the credential entry
-	var credKey [32]byte
-	copy(credKey[:20], issuerID[:])
-	copy(credKey[20:], subjectID[:12])
-
-	credKeylet := keylet.Keylet{Key: credKey, Type: 0x0081}
-
-	// Serialize credential data
-	credData := make([]byte, 64)
-	copy(credData[:20], issuerID[:])
-	copy(credData[20:40], subjectID[:])
-
-	if err := e.view.Insert(credKeylet, credData); err != nil {
-		return TefINTERNAL
-	}
-
-	account.OwnerCount++
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "Credential",
-		LedgerIndex:     hex.EncodeToString(credKey[:]),
-		NewFields: map[string]any{
-			"Issuer":         tx.Account,
-			"Subject":        tx.Subject,
-			"CredentialType": tx.CredentialType,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyCredentialAccept applies a CredentialAccept transaction
-func (e *Engine) applyCredentialAccept(tx *CredentialAccept, account *AccountRoot, metadata *Metadata) Result {
-	if tx.Issuer == "" || tx.CredentialType == "" {
-		return TemINVALID
-	}
-
-	issuerID, err := decodeAccountID(tx.Issuer)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	subjectID, _ := decodeAccountID(tx.Account)
-
-	// Find and update the credential
-	var credKey [32]byte
-	copy(credKey[:20], issuerID[:])
-	copy(credKey[20:], subjectID[:12])
-
-	credKeylet := keylet.Keylet{Key: credKey, Type: 0x0081}
-
-	_, err = e.view.Read(credKeylet)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "Credential",
-		LedgerIndex:     hex.EncodeToString(credKey[:]),
-		FinalFields: map[string]any{
-			"Accepted": true,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyCredentialDelete applies a CredentialDelete transaction
-func (e *Engine) applyCredentialDelete(tx *CredentialDelete, account *AccountRoot, metadata *Metadata) Result {
-	if tx.CredentialType == "" {
-		return TemINVALID
-	}
-
-	issuerID, _ := decodeAccountID(tx.Account)
-	var subjectID [20]byte
-	if tx.Subject != "" {
-		subjectID, _ = decodeAccountID(tx.Subject)
-	} else {
-		subjectID = issuerID
-	}
-
-	// Find and delete the credential
-	var credKey [32]byte
-	copy(credKey[:20], issuerID[:])
-	copy(credKey[20:], subjectID[:12])
-
-	credKeylet := keylet.Keylet{Key: credKey, Type: 0x0081}
-
-	if err := e.view.Erase(credKeylet); err != nil {
-		return TecNO_ENTRY
-	}
-
-	if account.OwnerCount > 0 {
-		account.OwnerCount--
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Credential",
-		LedgerIndex:     hex.EncodeToString(credKey[:]),
-	})
-
-	return TesSUCCESS
-}
-
-// applyPermissionedDomainSet applies a PermissionedDomainSet transaction
-func (e *Engine) applyPermissionedDomainSet(tx *PermissionedDomainSet, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	var domainKey [32]byte
-	if tx.DomainID != "" {
-		// Modifying existing domain
-		domainBytes, err := hex.DecodeString(tx.DomainID)
-		if err != nil || len(domainBytes) != 32 {
-			return TemINVALID
-		}
-		copy(domainKey[:], domainBytes)
-
-		domainKeylet := keylet.Keylet{Key: domainKey, Type: 0x0082}
-
-		_, err = e.view.Read(domainKeylet)
-		if err != nil {
-			return TecNO_ENTRY
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "ModifiedNode",
-			LedgerEntryType: "PermissionedDomain",
-			LedgerIndex:     tx.DomainID,
-		})
-	} else {
-		// Creating new domain
-		copy(domainKey[:20], accountID[:])
-		binary.BigEndian.PutUint32(domainKey[20:], account.Sequence)
-
-		domainKeylet := keylet.Keylet{Key: domainKey, Type: 0x0082}
-
-		domainData := make([]byte, 64)
-		copy(domainData[:20], accountID[:])
-
-		if err := e.view.Insert(domainKeylet, domainData); err != nil {
-			return TefINTERNAL
-		}
-
-		account.OwnerCount++
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "PermissionedDomain",
-			LedgerIndex:     hex.EncodeToString(domainKey[:]),
-			NewFields: map[string]any{
-				"Owner": tx.Account,
-			},
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// applyPermissionedDomainDelete applies a PermissionedDomainDelete transaction
-func (e *Engine) applyPermissionedDomainDelete(tx *PermissionedDomainDelete, account *AccountRoot, metadata *Metadata) Result {
-	if tx.DomainID == "" {
-		return TemINVALID
-	}
-
-	domainBytes, err := hex.DecodeString(tx.DomainID)
-	if err != nil || len(domainBytes) != 32 {
-		return TemINVALID
-	}
-
-	var domainKey [32]byte
-	copy(domainKey[:], domainBytes)
-
-	domainKeylet := keylet.Keylet{Key: domainKey, Type: 0x0082}
-
-	if err := e.view.Erase(domainKeylet); err != nil {
-		return TecNO_ENTRY
-	}
-
-	if account.OwnerCount > 0 {
-		account.OwnerCount--
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "PermissionedDomain",
-		LedgerIndex:     tx.DomainID,
-	})
-
-	return TesSUCCESS
-}
-
-// applyDelegateSet applies a DelegateSet transaction
-func (e *Engine) applyDelegateSet(tx *DelegateSet, account *AccountRoot, metadata *Metadata) Result {
-	accountID, _ := decodeAccountID(tx.Account)
-
-	if tx.Authorize != "" {
-		// Setting delegation
-		delegateID, err := decodeAccountID(tx.Authorize)
-		if err != nil {
-			return TecNO_TARGET
-		}
-
-		var delegateKey [32]byte
-		copy(delegateKey[:20], accountID[:])
-		copy(delegateKey[20:], delegateID[:12])
-
-		delegateKeylet := keylet.Keylet{Key: delegateKey, Type: 0x0083}
-
-		delegateData := make([]byte, 40)
-		copy(delegateData[:20], accountID[:])
-		copy(delegateData[20:40], delegateID[:])
-
-		if err := e.view.Insert(delegateKeylet, delegateData); err != nil {
-			// Try update if already exists
-			e.view.Update(delegateKeylet, delegateData)
-		}
-
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "Delegate",
-			LedgerIndex:     hex.EncodeToString(delegateKey[:]),
-			NewFields: map[string]any{
-				"Account":   tx.Account,
-				"Authorize": tx.Authorize,
-			},
-		})
-	}
-
-	return TesSUCCESS
-}
-
-// applyVaultCreate applies a VaultCreate transaction
-func (e *Engine) applyVaultCreate(tx *VaultCreate, account *AccountRoot, metadata *Metadata) Result {
-	if tx.Asset.Currency == "" {
-		return TemINVALID
-	}
-
-	accountID, _ := decodeAccountID(tx.Account)
-
-	// Create vault entry
-	var vaultKey [32]byte
-	copy(vaultKey[:20], accountID[:])
-	binary.BigEndian.PutUint32(vaultKey[20:], account.Sequence)
-
-	vaultKeylet := keylet.Keylet{Key: vaultKey, Type: 0x0084}
-
-	vaultData := make([]byte, 64)
-	copy(vaultData[:20], accountID[:])
-
-	if err := e.view.Insert(vaultKeylet, vaultData); err != nil {
-		return TefINTERNAL
-	}
-
-	account.OwnerCount++
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "CreatedNode",
-		LedgerEntryType: "Vault",
-		LedgerIndex:     hex.EncodeToString(vaultKey[:]),
-		NewFields: map[string]any{
-			"Owner": tx.Account,
-			"Asset": tx.Asset,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyVaultSet applies a VaultSet transaction
-func (e *Engine) applyVaultSet(tx *VaultSet, account *AccountRoot, metadata *Metadata) Result {
-	if tx.VaultID == "" {
-		return TemINVALID
-	}
-
-	vaultBytes, err := hex.DecodeString(tx.VaultID)
-	if err != nil || len(vaultBytes) != 32 {
-		return TemINVALID
-	}
-
-	var vaultKey [32]byte
-	copy(vaultKey[:], vaultBytes)
-
-	vaultKeylet := keylet.Keylet{Key: vaultKey, Type: 0x0084}
-
-	_, err = e.view.Read(vaultKeylet)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "Vault",
-		LedgerIndex:     tx.VaultID,
-	})
-
-	return TesSUCCESS
-}
-
-// applyVaultDelete applies a VaultDelete transaction
-func (e *Engine) applyVaultDelete(tx *VaultDelete, account *AccountRoot, metadata *Metadata) Result {
-	if tx.VaultID == "" {
-		return TemINVALID
-	}
-
-	vaultBytes, err := hex.DecodeString(tx.VaultID)
-	if err != nil || len(vaultBytes) != 32 {
-		return TemINVALID
-	}
-
-	var vaultKey [32]byte
-	copy(vaultKey[:], vaultBytes)
-
-	vaultKeylet := keylet.Keylet{Key: vaultKey, Type: 0x0084}
-
-	if err := e.view.Erase(vaultKeylet); err != nil {
-		return TecNO_ENTRY
-	}
-
-	if account.OwnerCount > 0 {
-		account.OwnerCount--
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "DeletedNode",
-		LedgerEntryType: "Vault",
-		LedgerIndex:     tx.VaultID,
-	})
-
-	return TesSUCCESS
-}
-
-// applyVaultDeposit applies a VaultDeposit transaction
-func (e *Engine) applyVaultDeposit(tx *VaultDeposit, account *AccountRoot, metadata *Metadata) Result {
-	if tx.VaultID == "" || tx.Amount.Value == "" {
-		return TemINVALID
-	}
-
-	vaultBytes, err := hex.DecodeString(tx.VaultID)
-	if err != nil || len(vaultBytes) != 32 {
-		return TemINVALID
-	}
-
-	var vaultKey [32]byte
-	copy(vaultKey[:], vaultBytes)
-
-	vaultKeylet := keylet.Keylet{Key: vaultKey, Type: 0x0084}
-
-	_, err = e.view.Read(vaultKeylet)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	// Deduct from account balance if XRP
-	if tx.Amount.Currency == "" || tx.Amount.Currency == "XRP" {
-		amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-		if err != nil {
-			return TemINVALID
-		}
-		if account.Balance < amount {
-			return TecINSUFFICIENT_FUNDS
-		}
-		account.Balance -= amount
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "Vault",
-		LedgerIndex:     tx.VaultID,
-		FinalFields: map[string]any{
-			"DepositAmount": tx.Amount,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyVaultWithdraw applies a VaultWithdraw transaction
-func (e *Engine) applyVaultWithdraw(tx *VaultWithdraw, account *AccountRoot, metadata *Metadata) Result {
-	if tx.VaultID == "" || tx.Amount.Value == "" {
-		return TemINVALID
-	}
-
-	vaultBytes, err := hex.DecodeString(tx.VaultID)
-	if err != nil || len(vaultBytes) != 32 {
-		return TemINVALID
-	}
-
-	var vaultKey [32]byte
-	copy(vaultKey[:], vaultBytes)
-
-	vaultKeylet := keylet.Keylet{Key: vaultKey, Type: 0x0084}
-
-	_, err = e.view.Read(vaultKeylet)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	// Add to account balance if XRP
-	if tx.Amount.Currency == "" || tx.Amount.Currency == "XRP" {
-		amount, err := strconv.ParseUint(tx.Amount.Value, 10, 64)
-		if err != nil {
-			return TemINVALID
-		}
-		account.Balance += amount
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "Vault",
-		LedgerIndex:     tx.VaultID,
-		FinalFields: map[string]any{
-			"WithdrawAmount": tx.Amount,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyVaultClawback applies a VaultClawback transaction
-func (e *Engine) applyVaultClawback(tx *VaultClawback, account *AccountRoot, metadata *Metadata) Result {
-	if tx.VaultID == "" || tx.Holder == "" {
-		return TemINVALID
-	}
-
-	vaultBytes, err := hex.DecodeString(tx.VaultID)
-	if err != nil || len(vaultBytes) != 32 {
-		return TemINVALID
-	}
-
-	var vaultKey [32]byte
-	copy(vaultKey[:], vaultBytes)
-
-	vaultKeylet := keylet.Keylet{Key: vaultKey, Type: 0x0084}
-
-	_, err = e.view.Read(vaultKeylet)
-	if err != nil {
-		return TecNO_ENTRY
-	}
-
-	_, err = decodeAccountID(tx.Holder)
-	if err != nil {
-		return TecNO_TARGET
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "Vault",
-		LedgerIndex:     tx.VaultID,
-		FinalFields: map[string]any{
-			"ClawbackHolder": tx.Holder,
-		},
-	})
-
-	return TesSUCCESS
-}
-
-// applyBatch applies a Batch transaction
-func (e *Engine) applyBatch(tx *Batch, account *AccountRoot, metadata *Metadata) Result {
-	if len(tx.RawTransactions) == 0 {
-		return TemINVALID
-	}
-
-	flags := tx.GetFlags()
-
-	// Process each raw transaction in the batch
-	for i, rawTx := range tx.RawTransactions {
-		// Decode and process the raw transaction blob
-		_, err := hex.DecodeString(rawTx.RawTransaction.RawTxBlob)
-		if err != nil {
-			if flags&BatchFlagAllOrNothing != 0 {
-				return TefINTERNAL
-			}
-			continue
-		}
-
-		// Record the batch processing
-		metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-			NodeType:        "CreatedNode",
-			LedgerEntryType: "BatchedTransaction",
-			NewFields: map[string]any{
-				"Index":    i,
-				"TxnBlob":  rawTx.RawTransaction.RawTxBlob,
-				"Executed": true,
-			},
-		})
-
-		// Check for early termination flags
-		if flags&BatchFlagUntilFailure != 0 {
-			// Would continue until a failure
-		}
-		if flags&BatchFlagOnlyOne != 0 {
-			// Only execute first successful one
-			break
-		}
-	}
-
-	return TesSUCCESS
-}
-
-// applyLedgerStateFix applies a LedgerStateFix transaction
-func (e *Engine) applyLedgerStateFix(tx *LedgerStateFix, account *AccountRoot, metadata *Metadata) Result {
-	// LedgerStateFix is a special admin transaction
-	// It can only be applied in certain conditions
-
-	if tx.Owner != "" {
-		_, err := decodeAccountID(tx.Owner)
-		if err != nil {
-			return TecNO_TARGET
-		}
-	}
-
-	metadata.AffectedNodes = append(metadata.AffectedNodes, AffectedNode{
-		NodeType:        "ModifiedNode",
-		LedgerEntryType: "LedgerStateFix",
-		NewFields: map[string]any{
-			"LedgerFixType": tx.LedgerFixType,
-			"Owner":         tx.Owner,
-		},
-	})
-
-	return TesSUCCESS
 }
